refactor: extract PDF page layout
