+++ src/research/pdf-regions.ts
@@ -17,7 +17,6 @@ import type {
   PdfRegionKind,
   PdfRegionLine,
   PdfSourceFragmentLineage,
-  PdfSourceSemanticFlowBoundaryDecision,
   PdfSourceRun,
 } from './import-types'
 import {
@@ -29,214 +28,89 @@ import {
   type PdfTextLine,
 } from './pdf-lines'
 import { copyPdfLinkedTokenSourceAnnotations } from './pdf-links'
+import {
+  classifyPdfFurniture,
+  firstPageTitleBlockRun,
+  furnitureRunKey,
+  marginTextIsNumeralOnly,
+  normalizeMarginText,
+  sourceRunBox,
+} from './pdf-furniture'
 import { normalizedNoteLabel } from './note-label'
 import { parsePdfScholarlyVisualLabel } from './pdf-scholarly-label'
-import { sanitizeXmlText } from './publication-integrity'
-import { sha256HexSync } from './sha256-sync'
+import {
+  buildReadingOrder,
+  inlineStackedFragmentOrder,
+  inlineStackedFragmentParts,
+  READING_ORDER_RESOLUTION_POLICY_VERSION,
+  READING_ORDER_RESOLUTION_THRESHOLD,
+  restoreInlineStackedAtomicUnits,
+  splitNoteDefinitionOrder,
+} from './pdf-reading-order'
+import {
+  marginBand,
+  proseDominantInlineMathLine,
+  sourceLineFlowOrder,
+  splitDetachedMathExtensionProseRuns,
+  splitSourceStackedInlineFormulaLines,
+  textLineFromRuns,
+  unresolvedMathExtensionLine,
+} from './pdf-source-math'
+import {
+  beginsVisualCaption,
+  classifiedLineFragment,
+  columnFor,
+  demoteInlineVisualReferenceContinuations,
+  detectColumns,
+  dominantIndent,
+  endsCaptionSentence,
+  endsIncompleteWrappedUrl,
+  markAdjacentPanelLabelContinuations,
+  promoteCaptionContinuations,
+  promoteNoteContinuations,
+  splitSourceCaptionLaneContinuations,
+  type ClassifiedLine,
+  type ColumnLayout,
+} from './pdf-page-layout'
+export {
+  classifyPageFurniture,
+  classifyPdfFurniture,
+  digitValue,
+  type PdfFurnitureAssessment,
+} from './pdf-furniture'
 
-type ClassifiedLine = PdfTextLine & {
-  id: string
-  kind: PdfRegionKind
-  confidence: number
-  noteLabel: string | null
-  captionContinuationSeedId?: string
-  panelLabelContinuationSeedId?: string
-  noteContinuationSeedId?: string
-  headingContinuationSeedId?: string
-  displayEquationClusterSeedId?: string
-  tabularGridBandId?: string
-  captionLaneSplitAmbiguous?: boolean
-  sourceFragmentLineage?: PdfSourceFragmentLineage
-  furniture?: PdfFurnitureEvidence
-  furnitureReview?: PdfFurnitureReview
-}
-
-type ColumnLayout = {
-  split: number | null
-  accepted: boolean
-  ambiguous: boolean
-  resolution: Omit<PdfReadingOrderResolution, 'page' | 'regionIds'> | null
-  tabularGridBands?: ReadonlyMap<string, string>
-  tabularGridColumns?: ReadonlyMap<string, PdfRegionColumn>
-}
+export {
+  canonicalPdfSourceSemanticFlowEvidence,
+  pdfBodySourceOrderExtremaByPage,
+  pdfSourceColumnFlowJoinOutcome,
+  pdfSourceColumnFlowStartsWithCjkNumericContinuation,
+  pdfSourceFragmentId,
+  pdfSourceSemanticFlowBoundaryDecisionId,
+  pdfSourceSemanticFlowRunSha256,
+  PDF_SOURCE_SEMANTIC_FLOW_BASE_EVIDENCE,
+  PDF_SOURCE_SEMANTIC_FLOW_COLUMN_EVIDENCE,
+  PDF_SOURCE_SEMANTIC_FLOW_CROSS_PAGE_EVIDENCE,
+  PDF_SOURCE_SEMANTIC_FLOW_NO_SPACE_EVIDENCE,
+  PDF_SOURCE_SEMANTIC_FLOW_SPACE_WHITESPACE_EVIDENCE,
+  type PdfBodySourceOrderExtremum,
+} from './pdf-source-semantic-flow'
+export {
+  proseDominantPdfMathSource,
+  sourceInlineFractionPairs,
+  sourceStackedMathPairs,
+} from './pdf-source-math'
+export { captionFontFamily } from './pdf-page-layout'
+export {
+  evaluateReadingOrder,
+  hasAcceptedCycle,
+  READING_ORDER_RESOLUTION_POLICY_VERSION,
+  READING_ORDER_RESOLUTION_THRESHOLD,
+} from './pdf-reading-order'
 
 const NOTE_LABEL = String.raw`(?:\d{1,3}|[*†‡§])`
-export const READING_ORDER_RESOLUTION_POLICY_VERSION = '1.0.0' as const
-export const READING_ORDER_RESOLUTION_THRESHOLD = 0.85
 
 function rounded(value: number) {
   return Math.round(value * 100_000) / 100_000
-}
-
-export function pdfSourceSemanticFlowRunSha256(run: PdfSourceRun) {
-  return sha256HexSync(
-    JSON.stringify([
-      run.page,
-      run.rotation,
-      run.method,
-      run.x,
-      run.y,
-      run.width,
-      run.height,
-      run.text.normalize('NFC'),
-      run.fontName,
-      run.fontSize,
-      run.sourceSequenceIndex ?? null,
-      run.sourceWhitespaceBefore ?? null,
-      run.sourceWhitespacePredecessorIndex ?? null,
-    ]),
-  )
-}
-
-export function pdfSourceFragmentId(line: PdfRegionLine) {
-  const lineage = line.sourceFragmentLineage
-  return lineage ? `${lineage.sourceLineId}:${lineage.fragment}` : null
-}
-
-export const PDF_SOURCE_SEMANTIC_FLOW_BASE_EVIDENCE = Object.freeze(
-  [
-    'exact-source-sequence-adjacency',
-    'font-baseline-compatible',
-    'explicit-fragment-lineage',
-  ].sort(),
-)
-
-export const PDF_SOURCE_SEMANTIC_FLOW_NO_SPACE_EVIDENCE = Object.freeze(
-  [
-    ...PDF_SOURCE_SEMANTIC_FLOW_BASE_EVIDENCE,
-    'continuation-punctuation',
-    'stacked-fragment-transition',
-  ].sort(),
-)
-
-export const PDF_SOURCE_SEMANTIC_FLOW_SPACE_WHITESPACE_EVIDENCE = Object.freeze(
-  [
-    ...PDF_SOURCE_SEMANTIC_FLOW_BASE_EVIDENCE,
-    'source-whitespace-separator',
-  ].sort(),
-)
-
-export const PDF_SOURCE_SEMANTIC_FLOW_COLUMN_EVIDENCE = Object.freeze(
-  [
-    'exact-source-sequence-adjacency',
-    'explicit-fragment-lineage',
-    'same-page-column-flow',
-    'same-page-column-geometry',
-  ].sort(),
-)
-
-// A page break is not a source-order adjacency: `sourceSequenceIndex` is the
-// per-page text-item index, so the tail of page N and the head of page N+1 are
-// never numerically adjacent. What the source can prove instead is that the
-// tail is the last prose text item its page paints and the continuation is the
-// first prose text item the next page paints, with only independently accounted
-// non-prose between them. That includes issue 042 page furniture and validated
-// visual/table ownership, neither of which may enter the paragraph.
-export const PDF_SOURCE_SEMANTIC_FLOW_CROSS_PAGE_EVIDENCE = Object.freeze(
-  [
-    'cross-page-column-geometry',
-    'explicit-fragment-lineage',
-    'non-prose-excluded-page-boundary',
-    'page-head-source-order-extremum',
-    'page-tail-source-order-extremum',
-  ].sort(),
-)
-
-export type PdfBodySourceOrderExtremum = { first: number; last: number }
-
-/**
- * Per-page first and last source text-item index over everything that is not
- * accounted page furniture or validated non-prose visual content. Extraction
- * and the independent quality audit both derive the cross-page prose boundary
- * from this map, so they cannot disagree about which runs a page break may skip.
- */
-export function pdfBodySourceOrderExtremaByPage(
-  regions: readonly Pick<
-    PdfPageRegion,
-    'id' | 'page' | 'lines' | 'furniture'
-  >[],
-  accountedNonProseRegionIds: ReadonlySet<string> = new Set(),
-) {
-  const extrema = new Map<number, PdfBodySourceOrderExtremum>()
-  for (const region of regions) {
-    if (region.furniture || accountedNonProseRegionIds.has(region.id)) continue
-    for (const line of region.lines) {
-      for (const run of line.runs) {
-        if (!run.text.trim() || run.sourceSequenceIndex === undefined) continue
-        const current = extrema.get(run.page)
-        if (!current) {
-          extrema.set(run.page, {
-            first: run.sourceSequenceIndex,
-            last: run.sourceSequenceIndex,
-          })
-          continue
-        }
-        current.first = Math.min(current.first, run.sourceSequenceIndex)
-        current.last = Math.max(current.last, run.sourceSequenceIndex)
-      }
-    }
-  }
-  return extrema
-}
-
-export function pdfSourceColumnFlowStartsWithCjkNumericContinuation(
-  continuationText: string,
-) {
-  return /^\p{N}+(?:[,.]\p{N}+)*(?:\p{Script=Han}|\p{Script=Hiragana}|\p{Script=Katakana})/u.test(
-    continuationText.trimStart(),
-  )
-}
-
-export function pdfSourceColumnFlowJoinOutcome(
-  language: string | null,
-  continuationText: string,
-  continuationRun: PdfSourceRun,
-) {
-  const hasSourceWhitespace =
-    continuationRun.sourceWhitespaceBefore === 'pdf-text-item' &&
-    continuationRun.sourceWhitespacePredecessorIndex !== undefined
-  if (hasSourceWhitespace) {
-    return { outcome: 'space' as const, separator: ' ' as const }
-  }
-  const cjkScript =
-    /^[^\p{L}\p{N}]*(?:\p{Script=Han}|\p{Script=Hiragana}|\p{Script=Katakana})/u.test(
-      continuationText,
-    ) || pdfSourceColumnFlowStartsWithCjkNumericContinuation(continuationText)
-  return cjkScript
-    ? { outcome: 'no-space' as const, separator: '' as const }
-    : { outcome: 'space' as const, separator: ' ' as const }
-}
-
-export function canonicalPdfSourceSemanticFlowEvidence(
-  evidence: readonly string[],
-) {
-  return [...new Set(evidence)].sort()
-}
-
-export function pdfSourceSemanticFlowBoundaryDecisionId(
-  decision: Omit<PdfSourceSemanticFlowBoundaryDecision, 'id'>,
-) {
-  const endpoint = (value: PdfSourceSemanticFlowBoundaryDecision['from']) => [
-    value.regionId,
-    value.lineId,
-    value.runIndex,
-    value.sourceSequenceIndex,
-    value.sourceRunSha256,
-    value.sourceFragmentId,
-  ]
-  return sha256HexSync(
-    JSON.stringify([
-      'pdf-source-semantic-flow-boundary-v1',
-      decision.page,
-      decision.rotation,
-      decision.method,
-      decision.topology,
-      decision.outcome,
-      endpoint(decision.from),
-      endpoint(decision.to),
-      canonicalPdfSourceSemanticFlowEvidence(decision.evidence),
-    ]),
-  )
 }
 
 function median(values: number[]) {
@@ -254,568 +128,6 @@ function quantile(values: number[], fraction: number) {
   return ordered[
     Math.min(ordered.length - 1, Math.floor(ordered.length * fraction))
   ]
-}
-
-function normalizeMarginText(text: string) {
-  return text
-    .toLocaleLowerCase()
-    .normalize('NFKC')
-    .replace(/[\p{N}]+/gu, '#')
-    .replace(/\b[ivxlcdm]+\b/giu, '#')
-    .replace(/\s+/g, ' ')
-    .trim()
-}
-
-type FurnitureRunOccurrence = {
-  run: PdfSourceRun
-  band: PdfFurnitureBand
-  key: string
-  normalizedText: string
-  numericValues: number[]
-}
-
-export type PdfFurnitureAssessment = {
-  patterns: Array<{
-    classification: PdfFurnitureClassification
-    band: PdfFurnitureBand
-    normalizedText: string
-    pages: number[]
-    boxes: NormalizedSourceBox[]
-    sequence?: number[]
-    evidence: string[]
-  }>
-  furnitureRuns: Array<{
-    run: PdfSourceRun
-    evidence: PdfFurnitureEvidence
-  }>
-  reviewRuns: Array<{
-    run: PdfSourceRun
-    review: PdfFurnitureReview
-  }>
-  repeatedMarginCount: number
-  byRunKey: ReadonlyMap<string, PdfFurnitureEvidence>
-  reviewByRunKey: ReadonlyMap<string, PdfFurnitureReview>
-  objectById: ReadonlyMap<string, PdfFurnitureEvidence>
-}
-
-function isQuarterTurn(rotation: number) {
-  const value = Math.abs(rotation % 180)
-  return Math.abs(value - 90) <= 2
-}
-
-function furnitureBand(
-  source: Pick<
-    NormalizedSourceBox,
-    'x' | 'y' | 'width' | 'height' | 'rotation'
-  >,
-): PdfFurnitureBand | null {
-  if (isQuarterTurn(source.rotation) && source.x <= 0.05) {
-    return 'left'
-  }
-  if (isQuarterTurn(source.rotation) && source.x + source.width >= 0.95) {
-    return 'right'
-  }
-  // Narrow, unrotated runs flush with the page edge are also margin bands.
-  // The width gate keeps ordinary two-column body text out of review.
-  if (source.x <= 0.05 && source.width <= 0.2) return 'left'
-  if (source.x + source.width >= 0.95 && source.width <= 0.2) return 'right'
-  if (source.y <= 0.1) return 'top'
-  if (source.y + source.height >= 0.9) return 'bottom'
-  return null
-}
-
-function sourceRunBox(run: PdfSourceRun): NormalizedSourceBox {
-  return {
-    page: run.page,
-    x: run.x,
-    y: run.y,
-    width: run.width,
-    height: run.height,
-    rotation: run.rotation,
-    method: run.method,
-  }
-}
-
-function furnitureRunKey(run: PdfSourceRun) {
-  return [
-    run.page,
-    run.text,
-    run.x,
-    run.y,
-    run.width,
-    run.height,
-    run.rotation,
-    run.fontName,
-    run.fontSize,
-    run.sourceSequenceIndex ?? '',
-  ].join('\u001f')
-}
-
-export function digitValue(character: string) {
-  const codePoint = character.codePointAt(0) ?? -1
-  const ranges = [
-    [0x30, 0x39],
-    [0x660, 0x669],
-    [0x6f0, 0x6f9],
-    [0x7c0, 0x7c9],
-    [0x966, 0x96f],
-    [0x9e6, 0x9ef],
-    [0xa66, 0xa6f],
-    [0xae6, 0xaef],
-    [0xb66, 0xb6f],
-    [0xbe6, 0xbef],
-    [0xc66, 0xc6f],
-    [0xce6, 0xcef],
-    [0xd66, 0xd6f],
-    [0xde6, 0xdef],
-    [0xe50, 0xe59],
-    [0xed0, 0xed9],
-    [0xf20, 0xf29],
-    [0x1040, 0x1049],
-    [0x1090, 0x1099],
-    [0x17e0, 0x17e9],
-    [0x1810, 0x1819],
-    [0x1946, 0x194f],
-    [0x19d0, 0x19d9],
-    [0x1a80, 0x1a89],
-    [0x1a90, 0x1a99],
-    [0x1b50, 0x1b59],
-    [0x1bb0, 0x1bb9],
-    [0x1c40, 0x1c49],
-    [0x1c50, 0x1c59],
-    [0xa620, 0xa629],
-    [0xa8d0, 0xa8d9],
-    [0xa900, 0xa909],
-    [0xa9d0, 0xa9d9],
-    [0xa9f0, 0xa9f9],
-    [0xaa50, 0xaa59],
-    [0xabf0, 0xabf9],
-    [0x104a0, 0x104a9],
-    [0x10d30, 0x10d39],
-    [0x10d40, 0x10d49],
-    [0x11066, 0x1106f],
-    [0x110f0, 0x110f9],
-    [0x11136, 0x1113f],
-    [0x111d0, 0x111d9],
-    [0x112f0, 0x112f9],
-    [0x11450, 0x11459],
-    [0x114d0, 0x114d9],
-    [0x11650, 0x11659],
-    [0x116c0, 0x116c9],
-    [0x116d0, 0x116d9],
-    [0x116da, 0x116e3],
-    [0x11730, 0x11739],
-    [0x118e0, 0x118e9],
-    [0x11950, 0x11959],
-    [0x11bf0, 0x11bf9],
-    [0x11c50, 0x11c59],
-    [0x11d50, 0x11d59],
-    [0x11da0, 0x11da9],
-    [0x11de0, 0x11de9],
-    [0x11f50, 0x11f59],
-    [0x16130, 0x16139],
-    [0x16a60, 0x16a69],
-    [0x16ac0, 0x16ac9],
-    [0x16b50, 0x16b59],
-    [0x16d70, 0x16d79],
-    [0x1ccf0, 0x1ccf9],
-    [0x1d7ce, 0x1d7ff],
-    [0x1e140, 0x1e149],
-    [0x1e2f0, 0x1e2f9],
-    [0x1e4f0, 0x1e4f9],
-    [0x1e5f1, 0x1e5fa],
-    [0x1e950, 0x1e959],
-    [0x1fbf0, 0x1fbf9],
-    [0xff10, 0xff19],
-  ] as const
-  for (const [start, end] of ranges) {
-    if (codePoint >= start && codePoint <= end) return codePoint - start
-  }
-  return null
-}
-
-function numericTokenValue(token: string) {
-  const digits = [...token]
-  if (
-    digits.length > 0 &&
-    digits.every((character) => digitValue(character) !== null)
-  ) {
-    return digits.reduce(
-      (value, character) => value * 10 + (digitValue(character) ?? 0),
-      0,
-    )
-  }
-  const roman = token.toLocaleLowerCase()
-  if (!/^[ivxlcdm]+$/u.test(roman)) return null
-  const values: Record<string, number> = {
-    i: 1,
-    v: 5,
-    x: 10,
-    l: 50,
-    c: 100,
-    d: 500,
-    m: 1000,
-  }
-  let total = 0
-  let previous = 0
-  for (const character of [...roman].reverse()) {
-    const value = values[character]
-    if (value < previous) total -= value
-    else {
-      total += value
-      previous = value
-    }
-  }
-  return total > 0 ? total : null
-}
-
-function marginNumericValues(text: string) {
-  const values: number[] = []
-  for (const match of text
-    .normalize('NFKC')
-    .matchAll(/[\p{N}]+|\b[ivxlcdm]+\b/giu)) {
-    const value = numericTokenValue(match[0])
-    if (value !== null) values.push(value)
-  }
-  return values
-}
-
-function marginTextIsNumeralOnly(text: string) {
-  return (
-    marginNumericValues(text).length > 0 &&
-    /^[\s.,:;()[\]{}+\-–—\p{N}ivxlcdm]+$/iu.test(text)
-  )
-}
-
-function firstPageTitleBlockRun(run: PdfSourceRun, bodyFontSize: number) {
-  if (isQuarterTurn(run.rotation)) return false
-  if (run.page !== 1 || run.y > 0.42) return false
-  if (run.y < 0.08 && run.fontSize < bodyFontSize * 1.2 && run.width < 0.55) {
-    return false
-  }
-  return (
-    run.y >= 0.08 || run.width >= 0.28 || run.fontSize >= bodyFontSize * 0.92
-  )
-}
-
-function furniturePatternProven(occurrences: FurnitureRunOccurrence[]) {
-  const pages = [
-    ...new Set(occurrences.map((occurrence) => occurrence.run.page)),
-  ].sort((left, right) => left - right)
-  if (pages.length < 2) return null
-  const hasNumbers = occurrences.some(
-    (occurrence) => occurrence.numericValues.length > 0,
-  )
-  const orderedOccurrences = occurrences
-    .slice()
-    .sort((left, right) => left.run.page - right.run.page)
-  const sequence = orderedOccurrences.flatMap((occurrence) => {
-    const values = occurrence.numericValues
-    return values.length > 0 ? [values.at(-1)!] : []
-  })
-  const stableNumericPrefix = orderedOccurrences.every((occurrence) => {
-    const values = occurrence.numericValues
-    const first = orderedOccurrences[0]?.numericValues ?? []
-    return (
-      values.length === first.length &&
-      values.slice(0, -1).every((value, index) => value === first[index])
-    )
-  })
-  const incrementing =
-    hasNumbers &&
-    sequence.length === pages.length &&
-    stableNumericPrefix &&
-    sequence.every(
-      (value, index) => index === 0 || value === sequence[index - 1] + 1,
-    )
-  const exactText =
-    new Set(
-      occurrences.map((occurrence) => occurrence.run.text.normalize('NFKC')),
-    ).size === 1
-  if (!exactText && !incrementing) return null
-  const rotated = occurrences.some((occurrence) =>
-    isQuarterTurn(occurrence.run.rotation),
-  )
-  const geometryKey = (occurrence: FurnitureRunOccurrence) =>
-    [
-      Math.round(occurrence.run.x / 0.025),
-      Math.round(occurrence.run.y / 0.025),
-      Math.round(occurrence.run.width / 0.025),
-      Math.round(occurrence.run.height / 0.01),
-      isQuarterTurn(occurrence.run.rotation)
-        ? Math.round(Math.abs(occurrence.run.rotation) / 90)
-        : 0,
-    ].join(':')
-  const stableGeometry = new Set(occurrences.map(geometryKey)).size === 1
-  const classification: PdfFurnitureClassification = rotated
-    ? 'rotated-margin'
-    : incrementing
-      ? 'incrementing-numeral'
-      : 'repeated-text'
-  const evidence = [
-    'cross-page-repetition',
-    'same-margin-band',
-    ...(stableGeometry
-      ? ['stable-source-geometry']
-      : ['bounded-margin-geometry']),
-    ...(incrementing ? ['incrementing-numeral-sequence'] : []),
-    ...(rotated ? ['quarter-turn-margin-rotation'] : []),
-  ]
-  return {
-    classification,
-    pages,
-    boxes: occurrences
-      .slice()
-      .sort(
-        (left, right) =>
-          left.run.page - right.run.page || left.run.x - right.run.x,
-      )
-      .map((occurrence) => sourceRunBox(occurrence.run)),
-    sequence: incrementing ? sequence : undefined,
-    evidence,
-  }
-}
-
-/**
- * Classify page furniture before line grouping.  The output keeps a stable
- * source-run key so line splitting can carry the proof into PdfPageRegion.
- */
-export function classifyPdfFurniture(
-  pages: readonly PdfPageAnalysis[],
-  {
-    deferredRunKeys = new Set<string>(),
-  }: { deferredRunKeys?: ReadonlySet<string> } = {},
-): PdfFurnitureAssessment {
-  const bodyFontSize = median(
-    pages
-      .flatMap((page) => page.runs.map((run) => run.fontSize))
-      .filter((size) => size > 0),
-  )
-  const occurrencesByPattern = new Map<string, FurnitureRunOccurrence[]>()
-  for (const page of pages) {
-    for (const run of page.runs) {
-      if (deferredRunKeys.has(furnitureRunKey(run))) continue
-      const band = furnitureBand(run)
-      if (!band || firstPageTitleBlockRun(run, bodyFontSize)) continue
-      const text = run.text.normalize('NFKC').replace(/\s+/gu, ' ').trim()
-      if (!text) continue
-      const numericValues = marginNumericValues(text)
-      const alphabeticCharacters = (text.match(/[\p{L}]/gu) ?? []).length
-      const numericOnly =
-        numericValues.length > 0 &&
-        (alphabeticCharacters === 0 || marginTextIsNumeralOnly(text))
-      if ((!numericOnly && text.length < 4) || run.width < 0.01) continue
-      const normalizedText = normalizeMarginText(text)
-      // The band, rather than an exact x coordinate, is the invariant across
-      // alternating-page headers and two-column folios.  Geometry is still
-      // retained in the evidence boxes and used to keep body text out of the
-      // candidate set.
-      const numericSystem =
-        numericValues.length > 0
-          ? (text
-              .match(/[\p{N}]+/u)?.[0]
-              ?.codePointAt(0)
-              ?.toString(16)
-              .slice(0, -1) ??
-            (marginTextIsNumeralOnly(text) ? 'roman' : 'numeric'))
-          : ''
-      const key = `${band}\u0000${normalizedText}\u0000${numericSystem}`
-      const list = occurrencesByPattern.get(key) ?? []
-      list.push({ run, band, key, normalizedText, numericValues })
-      occurrencesByPattern.set(key, list)
-    }
-  }
-  const furnitureRuns: Array<{
-    run: PdfSourceRun
-    evidence: PdfFurnitureEvidence
-  }> = []
-  const byRunKey = new Map<string, PdfFurnitureEvidence>()
-  const patterns: PdfFurnitureAssessment['patterns'] = []
-  for (const occurrences of occurrencesByPattern.values()) {
-    const proven = furniturePatternProven(occurrences)
-    if (!proven) continue
-    const first = occurrences[0]
-    const evidence: PdfFurnitureEvidence = {
-      classification: proven.classification,
-      band: first.band,
-      pages: proven.pages,
-      boxes: proven.boxes,
-      evidence: proven.evidence,
-      normalizedText: first.normalizedText,
-      ...(proven.sequence ? { sequence: proven.sequence } : {}),
-      sourceRunIndexes: occurrences
-        .map(({ run }) => run.sourceSequenceIndex)
-        .filter((index): index is number => index !== undefined)
-        .sort((left, right) => left - right),
-    }
-    patterns.push({
-      classification: proven.classification,
-      band: first.band,
-      normalizedText: first.normalizedText,
-      pages: proven.pages,
-      boxes: proven.boxes,
-      evidence: proven.evidence,
-      ...(proven.sequence ? { sequence: proven.sequence } : {}),
-    })
-    for (const occurrence of occurrences) {
-      const key = furnitureRunKey(occurrence.run)
-      byRunKey.set(key, evidence)
-      furnitureRuns.push({ run: occurrence.run, evidence })
-    }
-  }
-  // A quarter-turned margin stamp is self-proving from geometry and rotation;
-  // it does not need a text or journal-name match.  Keep this deliberately
-  // narrow so a single horizontal sentence at the bottom remains reviewable.
-  for (const page of pages) {
-    for (const run of page.runs) {
-      if (deferredRunKeys.has(furnitureRunKey(run))) continue
-      const band = furnitureBand(run)
-      if (
-        !band ||
-        !isQuarterTurn(run.rotation) ||
-        byRunKey.has(furnitureRunKey(run))
-      )
-        continue
-      const evidence: PdfFurnitureEvidence = {
-        classification: 'rotated-margin',
-        band,
-        pages: [run.page],
-        boxes: [sourceRunBox(run)],
-        evidence: ['quarter-turn-margin-rotation', 'margin-band-geometry'],
-        normalizedText: normalizeMarginText(run.text),
-        sourceRunIndexes:
-          run.sourceSequenceIndex === undefined
-            ? []
-            : [run.sourceSequenceIndex],
-      }
-      byRunKey.set(furnitureRunKey(run), evidence)
-      furnitureRuns.push({ run, evidence })
-    }
-  }
-  const reviewRuns: Array<{ run: PdfSourceRun; review: PdfFurnitureReview }> =
-    []
-  const reviewByRunKey = new Map<string, PdfFurnitureReview>()
-  for (const page of pages) {
-    for (const run of page.runs) {
-      const key = furnitureRunKey(run)
-      if (deferredRunKeys.has(key)) continue
-      if (byRunKey.has(key)) continue
-      const band = furnitureBand(run)
-      if (!band || firstPageTitleBlockRun(run, bodyFontSize)) continue
-      const edgeLike =
-        band === 'left' ||
-        band === 'right' ||
-        (band === 'top' && run.y <= 0.06) ||
-        (band === 'bottom' && run.y + run.height >= 0.92)
-      if (!edgeLike) continue
-      const text = run.text.replace(/\s+/gu, ' ').trim()
-      const words = text.match(/[\p{L}\p{N}]{2,}/gu) ?? []
-      if (text.length < 8 || words.length < 2 || run.width < 0.08) continue
-      const review: PdfFurnitureReview = {
-        reason: 'single-occurrence-margin',
-        band,
-        pages: [run.page],
-        boxes: [sourceRunBox(run)],
-        evidence: [
-          'single-page-occurrence',
-          'margin-band-geometry',
-          'repetition-unproven',
-        ],
-      }
-      reviewByRunKey.set(key, review)
-      reviewRuns.push({ run, review })
-    }
-  }
-  const objectById = new Map<string, PdfFurnitureEvidence>()
-  const ruleGroups = new Map<
-    string,
-    Array<{
-      object: NonNullable<PdfPageAnalysis['objects']>[number]
-      band: PdfFurnitureBand
-    }>
-  >()
-  for (const page of pages) {
-    for (const object of page.objects ?? []) {
-      const band = furnitureBand(object.box)
-      if (
-        !band ||
-        object.kind !== 'vector' ||
-        object.box.width < 0.45 ||
-        object.box.height > 0.015
-      ) {
-        continue
-      }
-      const key = [
-        band,
-        Math.round(object.box.x / 0.025),
-        Math.round(object.box.y / 0.025),
-        Math.round(object.box.width / 0.025),
-        Math.round(object.box.height / 0.01),
-      ].join(':')
-      const group = ruleGroups.get(key) ?? []
-      group.push({ object, band })
-      ruleGroups.set(key, group)
-    }
-  }
-  const hasMarginRunOnPage = new Set(furnitureRuns.map(({ run }) => run.page))
-  for (const group of ruleGroups.values()) {
-    const pagesForRule = [
-      ...new Set(group.map(({ object }) => object.page)),
-    ].sort((left, right) => left - right)
-    if (
-      pagesForRule.length < 2 &&
-      !pagesForRule.some((page) => hasMarginRunOnPage.has(page))
-    ) {
-      continue
-    }
-    const first = group[0]
-    const evidence: PdfFurnitureEvidence = {
-      classification: 'separator-rule',
-      band: first.band,
-      pages: pagesForRule,
-      boxes: group
-        .slice()
-        .sort((left, right) => left.object.page - right.object.page)
-        .map(({ object }) => ({ ...object.box })),
-      evidence: [
-        ...(pagesForRule.length >= 2 ? ['cross-page-repetition'] : []),
-        'same-margin-band',
-        'thin-rule-geometry',
-      ],
-    }
-    for (const { object } of group) objectById.set(object.id, evidence)
-  }
-  return {
-    patterns: patterns.sort(
-      (left, right) =>
-        left.pages[0] - right.pages[0] ||
-        left.band.localeCompare(right.band) ||
-        left.normalizedText.localeCompare(right.normalizedText),
-    ),
-    furnitureRuns,
-    reviewRuns,
-    repeatedMarginCount: patterns.length,
-    byRunKey,
-    reviewByRunKey,
-    objectById,
-  }
-}
-
-/** Backwards-compatible descriptive alias for callers that name the stage. */
-export const classifyPageFurniture = classifyPdfFurniture
-
-function beginsVisualCaption(text: string) {
-  const normalized = text.replace(/\s+/gu, ' ').trim()
-  return (
-    Boolean(
-      parsePdfScholarlyVisualLabel(normalized, {
-        context: 'caption',
-      }),
-    ) ||
-    /^[^.!?;:]{1,72}\s+(?:fig(?:ure)?|table)\.?\s*(?:\d+|[ivxlcdm]+)\s*[.:–—-]\s*\p{Lu}/iu.test(
-      normalized,
-    )
-  )
 }
 
 function beginsSourceStyledVisualCaption(line: PdfTextLine) {
@@ -845,857 +157,6 @@ function beginsSourceStyledVisualCaption(line: PdfTextLine) {
     styledLabel.identifier === lineLabel.identifier &&
     styledLabel.consumedEnd === styledPrefix.length
   )
-}
-
-function unpublishableText(text: string) {
-  return text.includes('\ufffd') || sanitizeXmlText(text) !== text
-}
-
-function unresolvedMathExtensionLine(line: PdfTextLine) {
-  return line.runs.some(
-    (run) => /CMEX\d*/iu.test(run.fontName) && unpublishableText(run.text),
-  )
-}
-
-export function proseDominantPdfMathSource(source: {
-  text: string
-  width: number
-  runs: readonly PdfSourceRun[]
-}) {
-  const text = source.text.replace(/\s+/gu, ' ').trim()
-  const proseWords = text.match(/\p{L}{3,}/gu) ?? []
-  const visibleRuns = source.runs.filter((run) => run.text.trim())
-  const visibleCharacters = visibleRuns.reduce(
-    (total, run) => total + run.text.replace(/\s/gu, '').length,
-    0,
-  )
-  const mathCharacters = visibleRuns.reduce(
-    (total, run) =>
-      total +
-      (/(?:cmmi|cmsy|cmex|math|symbol)/iu.test(run.fontName)
-        ? run.text.replace(/\s/gu, '').length
-        : 0),
-    0,
-  )
-  const proseFontShare =
-    visibleCharacters > 0 ? 1 - mathCharacters / visibleCharacters : 1
-  const numberedInstruction =
-    /^(?:\d+|[A-Za-z])[.)]\s+\p{Lu}/u.test(text) && proseWords.length >= 3
-  const relationCount = (text.match(/(?:<=|>=|==|!=|=|≤|≥|≈|≠|<|>)/gu) ?? [])
-    .length
-  const alphabeticTokenCount = (text.match(/\p{L}{2,}/gu) ?? []).length
-  const strongMathSignal =
-    /[\p{Script=Greek}\p{N}∆_=+*/<>^−×÷≤≥≈≠∼⊙∂∞∏∈∉→←∫∑√]/u.test(text)
-  const openingDelimiter = /[\p{Ps}\p{Pi}]/u
-  const closingDelimiter = /[\p{Pe}\p{Pf}]/u
-  const delimiterDepth = (value: string) =>
-    Array.from(value).reduce(
-      (depth, character) =>
-        depth +
-        Number(openingDelimiter.test(character)) -
-        Number(closingDelimiter.test(character)),
-      0,
-    )
-  const insideSourceDelimiter = (start: number, end: number) => {
-    const before = text.slice(0, start)
-    const after = text.slice(end)
-    const pairedStraightQuote = ['"', "'"].some(
-      (quote) => before.lastIndexOf(quote) >= 0 && after.indexOf(quote) >= 0,
-    )
-    return (
-      delimiterDepth(before) > 0 ||
-      delimiterDepth(after) < 0 ||
-      pairedStraightQuote
-    )
-  }
-  const compactMathContext = /[\p{Script=Greek}∆_=+*/<>^−×÷≤≥≈≠∼⊙∂∞∏∈∉→←∫∑√]/u
-  const compactLeftAtom = /[\p{L}\p{N}\p{Pe}]/u
-  const compactRightAtom = /[\p{L}\p{N}\p{Ps}]/u
-  const delimitedUnsupportedAlphabeticTokens = [
-    ...text.matchAll(/\p{L}{2,}/gu),
-  ].flatMap((match) => {
-    const word = match[0]
-    const start = match.index
-    const end = start + word.length
-    const compactBefore =
-      compactMathContext.test(text[start - 1] ?? '') &&
-      compactLeftAtom.test(text[start - 2] ?? '')
-    const compactAfter =
-      compactMathContext.test(text[end] ?? '') &&
-      compactRightAtom.test(text[end + 1] ?? '')
-    const compactSymbolIdentifier =
-      Array.from(word).length <= 3 && (compactBefore || compactAfter)
-    return insideSourceDelimiter(start, end) &&
-      !/^\p{Ll}\p{Lu}$/u.test(word) &&
-      !/^d(?:\p{Ll}|\p{Script=Greek}){1,2}$/u.test(word) &&
-      !/\p{Script=Greek}/u.test(word) &&
-      !compactSymbolIdentifier &&
-      !/^(?:arg|cosh?|det|diag|dim|exp|gcd|lim|log|max|min|mod|sinh?|sqrt|tanh?|var)$/iu.test(
-        word,
-      )
-      ? [{ word, start, end }]
-      : []
-  })
-  const unresolvedDelimitedProse =
-    unpublishableText(text) &&
-    (delimitedUnsupportedAlphabeticTokens.length >= 2 ||
-      delimitedUnsupportedAlphabeticTokens.some(({ end }) =>
-        /^[^\p{L}]{0,8}[.!?]/u.test(text.slice(end)),
-      ))
-  const shortWordSentence =
-    alphabeticTokenCount >= 3 &&
-    relationCount === 0 &&
-    (!strongMathSignal || /[.!?](?:\s|$)/u.test(text))
-  const narrowProseSentence =
-    source.width < 0.3 &&
-    proseWords.length >= 4 &&
-    proseFontShare >= 0.7 &&
-    relationCount <= 1 &&
-    /(?:[.!?]\s*$|[,;:](?:\s|$)|\s[-–—]\s)/u.test(text) &&
-    /\b(?:this|that|these|those|is|are|was|were|be|been|being|has|have|had|with|without|not|remains?|after|before|because|while|where|which|who|we|our|their)\b/iu.test(
-      text,
-    )
-  const narrowProseLeadIn =
-    source.width < 0.3 &&
-    proseFontShare >= 0.7 &&
-    relationCount === 0 &&
-    alphabeticTokenCount >= 1 &&
-    text.includes(':')
-  return (
-    numberedInstruction ||
-    unresolvedDelimitedProse ||
-    shortWordSentence ||
-    narrowProseLeadIn ||
-    narrowProseSentence ||
-    (source.width >= 0.3 && proseWords.length >= 4 && proseFontShare >= 0.5)
-  )
-}
-
-function proseDominantInlineMathLine(line: PdfTextLine) {
-  return proseDominantPdfMathSource(line)
-}
-
-function marginBand(
-  source: Pick<PdfTextLine, 'x' | 'y' | 'width' | 'height'> & {
-    rotation?: number
-  },
-) {
-  return (
-    source.y <= 0.1 ||
-    source.y + source.height >= 0.9 ||
-    (isQuarterTurn(source.rotation ?? 0) &&
-      (source.x <= 0.12 || source.x + source.width >= 0.88))
-  )
-}
-
-function textLineFromRuns(source: PdfTextLine, runs: PdfTextLine['runs']) {
-  const left = Math.min(...runs.map((run) => run.x))
-  const top = Math.min(...runs.map((run) => run.y))
-  const right = Math.max(...runs.map((run) => run.x + run.width))
-  const bottom = Math.max(...runs.map((run) => run.y + run.height))
-  const fragment = {
-    ...source,
-    text: mergePdfRunText(runs),
-    x: left,
-    y: top,
-    width: right - left,
-    height: bottom - top,
-    fontSize: Math.max(...runs.map((run) => run.fontSize)),
-    runs,
-  }
-  copyPdfLinkedTokenSourceAnnotations(source, fragment)
-  return fragment
-}
-
-function computerModernMathSourceRun(
-  run: PdfTextLine['runs'][number],
-  maximumFontSize: number,
-) {
-  const text = run.text.replace(/\s+/gu, ' ').trim()
-  if (!text) return false
-  if (/(?:^|[+_-])CM(?:MI|SY|EX)\d*(?:$|[+_-])/iu.test(run.fontName)) {
-    return true
-  }
-  if (!/(?:^|[+_-])CMR\d*(?:$|[+_-])/iu.test(run.fontName)) return false
-  const proseWords = text.match(/\p{L}{3,}/gu) ?? []
-  return (
-    run.fontSize <= maximumFontSize * 0.82 ||
-    proseWords.length === 0 ||
-    (proseWords.length === 1 &&
-      /^(?:arg|cosh?|diag|exp|min|max|log|sinh?|sqrt|tanh?|var)\b/iu.test(text))
-  )
-}
-
-function horizontallyStackedMathPair(
-  left: PdfTextLine['runs'][number],
-  right: PdfTextLine['runs'][number],
-  maximumFontSize: number,
-) {
-  if (
-    left.fontSize > maximumFontSize * 0.86 ||
-    right.fontSize > maximumFontSize * 0.86
-  ) {
-    return false
-  }
-  const horizontalOverlap = Math.max(
-    0,
-    Math.min(left.x + left.width, right.x + right.width) -
-      Math.max(left.x, right.x),
-  )
-  const minimumWidth = Math.min(left.width, right.width)
-  if (
-    minimumWidth <= 0 ||
-    horizontalOverlap < minimumWidth * 0.7 ||
-    Math.abs(left.x + left.width / 2 - (right.x + right.width / 2)) >
-      Math.max(0.008, Math.max(left.width, right.width) * 0.3)
-  ) {
-    return false
-  }
-  const leftCenterY = left.y + left.height / 2
-  const rightCenterY = right.y + right.height / 2
-  const minimumHeight = Math.min(left.height, right.height)
-  const verticalGap = Math.max(
-    left.y - (right.y + right.height),
-    right.y - (left.y + left.height),
-    0,
-  )
-  return (
-    Math.abs(leftCenterY - rightCenterY) >=
-      Math.max(0.003, minimumHeight * 0.45) &&
-    verticalGap <= Math.max(0.006, minimumHeight * 0.7)
-  )
-}
-
-function horizontalBoxGap(
-  left: PdfTextLine['runs'][number],
-  right: PdfTextLine['runs'][number],
-) {
-  return Math.max(
-    left.x - (right.x + right.width),
-    right.x - (left.x + left.width),
-    0,
-  )
-}
-
-export function sourceStackedMathPairs(
-  runs: PdfTextLine['runs'],
-  maximumFontSize: number,
-) {
-  return runs.flatMap((left, leftIndex) =>
-    runs
-      .slice(leftIndex + 1)
-      .flatMap((right) =>
-        horizontallyStackedMathPair(left, right, maximumFontSize)
-          ? [[left, right] as const]
-          : [],
-      ),
-  )
-}
-
-export function sourceInlineFractionPairs(line: PdfTextLine) {
-  const runs = line.runs.filter((run) => run.text.trim())
-  if (runs.length < 2 || runs.some(mathExtensionSourceRun)) return []
-  const maximumFontSize = Math.max(...runs.map((run) => run.fontSize))
-  // PDF placement alone cannot distinguish a stacked numerator/denominator
-  // from simultaneous super/subscripts, including mixed letter pairs such as
-  // x/y beside a coefficient. Preserve every such pair as a two-dimensional
-  // equation obligation instead of guessing canonical inline semantics.
-  return sourceStackedMathPairs(runs, maximumFontSize)
-}
-
-function verticalLineOverlap(left: PdfTextLine, right: PdfTextLine) {
-  return Math.max(
-    0,
-    Math.min(left.y + left.height, right.y + right.height) -
-      Math.max(left.y, right.y),
-  )
-}
-
-function mathExtensionSourceRun(run: PdfTextLine['runs'][number]) {
-  return (
-    /(?:^|[+_-])CMEX\d*(?:$|[+_-])/iu.test(run.fontName) ||
-    /^[∑∫∏√]$/u.test(run.text.trim())
-  )
-}
-
-/**
- * Large operators can cause PDF.js to split one visual row into two source
- * lines: the operator and its scripts land in one line while a neighboring
- * fraction plus prose land in another. Coalesce only the proved source case:
- * a vertically interleaved math-extension line beside a companion that owns a
- * small stacked pair and prose before the operator. The ordinary inline
- * splitter can then retain prose and crop the complete 2-D formula envelope.
- */
-function mergeInterleavedStackedFormulaLines(lines: PdfTextLine[]) {
-  const claimed = new Set<number>()
-  const mergedAt = new Map<number, PdfTextLine>()
-
-  for (const [lineIndex, line] of lines.entries()) {
-    if (claimed.has(lineIndex)) continue
-    const extensionRuns = line.runs.filter(mathExtensionSourceRun)
-    if (extensionRuns.length === 0) continue
-    const extensionLeft = Math.min(...extensionRuns.map((run) => run.x))
-    const candidates = lines
-      .map((candidate, candidateIndex) => ({ candidate, candidateIndex }))
-      .filter(({ candidate, candidateIndex }) => {
-        if (
-          candidateIndex === lineIndex ||
-          claimed.has(candidateIndex) ||
-          candidate.page !== line.page
-        ) {
-          return false
-        }
-        const overlap = verticalLineOverlap(line, candidate)
-        if (overlap < Math.min(line.height, candidate.height) * 0.45) {
-          return false
-        }
-        const maximumFontSize = Math.max(
-          ...candidate.runs.map((run) => run.fontSize),
-        )
-        const pairs = sourceStackedMathPairs(candidate.runs, maximumFontSize)
-        if (pairs.length === 0) return false
-        const pairNearExtension = pairs.some((pair) =>
-          pair.some((run) =>
-            extensionRuns.some(
-              (extension) => horizontalBoxGap(run, extension) <= 0.02,
-            ),
-          ),
-        )
-        if (!pairNearExtension) return false
-        return candidate.runs.some(
-          (run) =>
-            run.x + run.width <= extensionLeft &&
-            !computerModernMathSourceRun(run, maximumFontSize) &&
-            (run.text.match(/\p{L}{2,}/gu)?.length ?? 0) > 0,
-        )
-      })
-      .sort(
-        (left, right) =>
-          Math.abs(left.candidate.y - line.y) -
-            Math.abs(right.candidate.y - line.y) ||
-          left.candidateIndex - right.candidateIndex,
-      )
-    const match = candidates[0]
-    if (!match) continue
-
-    const runs = [...line.runs, ...match.candidate.runs].sort(
-      (left, right) => left.x - right.x || left.y - right.y,
-    )
-    const merged = textLineFromRuns(line, runs)
-    const outputIndex = Math.min(lineIndex, match.candidateIndex)
-    mergedAt.set(outputIndex, merged)
-    claimed.add(lineIndex)
-    claimed.add(match.candidateIndex)
-  }
-
-  return lines.flatMap((line, lineIndex) => {
-    const merged = mergedAt.get(lineIndex)
-    if (merged) return [merged]
-    return claimed.has(lineIndex) ? [] : [line]
-  })
-}
-
-/**
- * A PDF text line can contain a prose prefix and a visually stacked fraction
- * or simultaneous super/subscripts. Linear text merging cannot distinguish
- * those semantics. Split the source-backed two-dimensional run geometry,
- * retaining the prose runs as ordinary text and the exact run envelope as an
- * equation crop candidate.
- */
-function splitSourceStackedInlineFormulaLines(lines: PdfTextLine[]) {
-  return mergeInterleavedStackedFormulaLines(lines).flatMap(
-    (line, lineIndex) => {
-      const runs = line.runs.filter((run) => run.text.trim())
-      if (runs.length < 2) return [line]
-      const maximumFontSize = Math.max(...runs.map((run) => run.fontSize))
-      const stackedPairs = sourceStackedMathPairs(runs, maximumFontSize)
-      if (stackedPairs.length === 0) return [line]
-      const baseId = `page-${String(line.page).padStart(3, '0')}-inline-stacked-${String(lineIndex + 1).padStart(4, '0')}`
-      const withFragmentLineage = (
-        fragmentLine: PdfTextLine,
-        part: 'before' | 'formula' | 'after',
-      ) => {
-        const visibleRuns = fragmentLine.runs.filter((run) => run.text.trim())
-        const sourceSequenceIndexes = visibleRuns.flatMap((run) =>
-          run.sourceSequenceIndex === undefined
-            ? []
-            : [run.sourceSequenceIndex],
-        )
-        const sourceFragmentLineage =
-          sourceSequenceIndexes.length === visibleRuns.length &&
-          new Set(sourceSequenceIndexes).size === sourceSequenceIndexes.length
-            ? {
-                algorithm: 'source-run-fragment-v1' as const,
-                sourceLineId: line.id ?? baseId,
-                fragment: `inline-stacked-${part}` as const,
-                sourceSequenceIndexes,
-              }
-            : undefined
-        return {
-          ...fragmentLine,
-          ...(sourceFragmentLineage ? { sourceFragmentLineage } : {}),
-        }
-      }
-
-      const formulaRuns = new Set(stackedPairs.flat())
-      let expanded = true
-      while (expanded) {
-        expanded = false
-        for (const candidate of runs) {
-          const compactCandidateText = candidate.text.replace(/\s+/gu, '')
-          const adjacentBaselineAnchor =
-            candidate.fontSize >= maximumFontSize * 0.9 &&
-            /^[\p{L}\p{N})\]}]{1,4}$/u.test(compactCandidateText) &&
-            stackedPairs.some((pair) => {
-              const pairLeft = Math.min(...pair.map((run) => run.x))
-              const pairCenters = pair
-                .map((run) => run.y + run.height / 2)
-                .sort((left, right) => left - right)
-              const horizontalGap = pairLeft - (candidate.x + candidate.width)
-              const candidateCenter = candidate.y + candidate.height / 2
-              return (
-                !pair.includes(candidate) &&
-                horizontalGap >= -0.001 &&
-                horizontalGap <= Math.max(0.004, candidate.width * 0.45) &&
-                candidateCenter > pairCenters[0] &&
-                candidateCenter < pairCenters[1]
-              )
-            })
-          if (
-            formulaRuns.has(candidate) ||
-            (!computerModernMathSourceRun(candidate, maximumFontSize) &&
-              !adjacentBaselineAnchor)
-          ) {
-            continue
-          }
-          if (
-            [...formulaRuns].some(
-              (owned) => horizontalBoxGap(owned, candidate) <= 0.02,
-            )
-          ) {
-            formulaRuns.add(candidate)
-            expanded = true
-          }
-        }
-      }
-      const formulaLeft = Math.min(...[...formulaRuns].map((run) => run.x))
-      const formulaRight = Math.max(
-        ...[...formulaRuns].map((run) => run.x + run.width),
-      )
-      for (const candidate of runs) {
-        if (
-          !formulaRuns.has(candidate) &&
-          /^[,.;:!?()[\]{}]+$/u.test(candidate.text.trim()) &&
-          horizontalBoxGap(candidate, {
-            ...candidate,
-            x: formulaLeft,
-            width: formulaRight - formulaLeft,
-          }) <= 0.008
-        ) {
-          formulaRuns.add(candidate)
-        }
-      }
-      const beforeRuns = runs.filter(
-        (run) => !formulaRuns.has(run) && run.x + run.width <= formulaLeft,
-      )
-      const afterRuns = runs.filter(
-        (run) => !formulaRuns.has(run) && run.x >= formulaRight,
-      )
-      const unownedRuns = runs.filter(
-        (run) =>
-          !formulaRuns.has(run) &&
-          !beforeRuns.includes(run) &&
-          !afterRuns.includes(run),
-      )
-      const proseWordCount =
-        [...beforeRuns, ...afterRuns]
-          .map((run) => run.text)
-          .join(' ')
-          .match(/\p{L}{2,}/gu)?.length ?? 0
-      const hasProseSiblings = beforeRuns.length > 0 || afterRuns.length > 0
-      if (unownedRuns.length > 0 || (hasProseSiblings && proseWordCount < 2)) {
-        // The stacked source geometry is still ambiguous even when the
-        // surrounding text cannot be split conservatively. Withhold the whole
-        // line as one equation obligation rather than publishing guessed
-        // super/subscript semantics.
-        return [
-          withFragmentLineage({ ...line, id: `${baseId}-formula` }, 'formula'),
-        ]
-      }
-
-      const fragments: PdfTextLine[] = []
-      if (beforeRuns.length > 0) {
-        fragments.push(
-          withFragmentLineage(
-            {
-              ...textLineFromRuns(line, beforeRuns),
-              id: `${baseId}-before`,
-            },
-            'before',
-          ),
-        )
-      }
-      fragments.push(
-        withFragmentLineage(
-          {
-            ...textLineFromRuns(
-              line,
-              [...formulaRuns].sort(
-                (left, right) => left.x - right.x || left.y - right.y,
-              ),
-            ),
-            id: `${baseId}-formula`,
-          },
-          'formula',
-        ),
-      )
-      if (afterRuns.length > 0) {
-        fragments.push(
-          withFragmentLineage(
-            {
-              ...textLineFromRuns(line, afterRuns),
-              id: `${baseId}-after`,
-            },
-            'after',
-          ),
-        )
-      }
-      return fragments
-    },
-  )
-}
-
-function splitDetachedMathExtensionProseRuns(lines: PdfTextLine[]) {
-  const mathFont = (fontName: string) =>
-    /(?:cmmi|cmsy|cmex|math|symbol)/iu.test(fontName)
-  const extensionFont = (fontName: string) =>
-    /(?:CMEX|MathExtensions)/iu.test(fontName)
-  const sourceLineId = (line: PdfTextLine, lineIndex: number) =>
-    line.id ??
-    `page-${String(line.page).padStart(3, '0')}-source-line-${String(lineIndex + 1).padStart(4, '0')}`
-  const lineIndexes = new Map(lines.map((line, lineIndex) => [line, lineIndex]))
-  const exactRadicalHosts = new Map<PdfSourceRun, string>()
-  const exactHostLineIds = new Map<PdfTextLine, string>()
-  const runEntries = lines.flatMap((line) =>
-    line.runs.filter((run) => run.text.trim()).map((run) => ({ line, run })),
-  )
-
-  for (const { line: rootLine, run: root } of runEntries) {
-    if (
-      root.text.trim() !== '√' ||
-      !extensionFont(root.fontName) ||
-      root.sourceSequenceIndex === undefined
-    ) {
-      continue
-    }
-    const candidates = runEntries.filter(({ line, run: candidate }) => {
-      if (
-        line === rootLine ||
-        line.page !== rootLine.page ||
-        candidate.sourceSequenceIndex !== root.sourceSequenceIndex! + 1 ||
-        !candidate.text.trim() ||
-        extensionFont(candidate.fontName) ||
-        !mathFont(candidate.fontName)
-      ) {
-        return false
-      }
-      if (
-        root.sourceTextPaint &&
-        candidate.sourceTextPaint &&
-        (root.sourceTextPaint.textLedgerSha256 !==
-          candidate.sourceTextPaint.textLedgerSha256 ||
-          root.sourceTextPaint.operatorLedgerSha256 !==
-            candidate.sourceTextPaint.operatorLedgerSha256)
-      ) {
-        return false
-      }
-      const horizontalTolerance = Math.max(0.00001, root.width * 0.001)
-      const radicalBottom = root.y + root.height
-      const verticalTolerance = Math.max(0.001, root.height * 0.08)
-      const fontSizeRatio = candidate.fontSize / root.fontSize
-      const heightRatio = candidate.height / root.height
-      return (
-        Math.abs(candidate.x - (root.x + root.width)) <= horizontalTolerance &&
-        candidate.y > root.y + root.height * 0.5 &&
-        Math.abs(candidate.y - radicalBottom) <= verticalTolerance &&
-        fontSizeRatio >= 0.65 &&
-        fontSizeRatio <= 1.5 &&
-        heightRatio >= 0.65 &&
-        heightRatio <= 1.5
-      )
-    })
-    if (candidates.length !== 1) continue
-    const hostLine = candidates[0].line
-    const hostLineIndex = lineIndexes.get(hostLine)
-    if (hostLineIndex === undefined) continue
-    const hostLineId = sourceLineId(hostLine, hostLineIndex)
-    exactHostLineIds.set(hostLine, hostLineId)
-    exactRadicalHosts.set(root, hostLineId)
-  }
-  const probableHost = (line: PdfTextLine) => {
-    if (proseDominantInlineMathLine(line)) return false
-    if (/-inline-stacked-\d+-formula$/u.test(line.id ?? '')) return true
-    const runs = line.runs.filter((run) => run.text.trim())
-    const visibleCharacters = runs.reduce(
-      (total, run) => total + run.text.replace(/\s+/gu, '').length,
-      0,
-    )
-    const mathCharacters = runs
-      .filter((run) => mathFont(run.fontName))
-      .reduce((total, run) => total + run.text.replace(/\s+/gu, '').length, 0)
-    return (
-      visibleCharacters > 0 &&
-      mathCharacters / visibleCharacters >= 0.7 &&
-      /[=+−×÷∫∑√≤≥≈]/u.test(line.text)
-    )
-  }
-  const hosts = lines.filter(probableHost)
-  const boxDistance = (run: PdfSourceRun, line: PdfTextLine) => {
-    const horizontalGap = Math.max(
-      line.x - (run.x + run.width),
-      run.x - (line.x + line.width),
-      0,
-    )
-    const verticalGap = Math.max(
-      line.y - (run.y + run.height),
-      run.y - (line.y + line.height),
-      0,
-    )
-    const centerDistance = Math.abs(
-      run.x + run.width / 2 - (line.x + line.width / 2),
-    )
-    return verticalGap * 2 + horizontalGap + centerDistance * 0.05
-  }
-
-  return lines.flatMap((line, lineIndex) => {
-    const preparedLine =
-      !line.id && exactHostLineIds.has(line)
-        ? { ...line, id: exactHostLineIds.get(line)! }
-        : line
-    const runs = line.runs.filter((run) => run.text.trim())
-    const extensionRuns = runs.filter(
-      (run) =>
-        extensionFont(run.fontName) &&
-        /^[√∫∑∏()[\]{}|]+$/u.test(run.text.trim()),
-    )
-    const proseRuns = runs.filter((run) => !extensionRuns.includes(run))
-    if (extensionRuns.length === 0) return [preparedLine]
-    const fallbackProseCue =
-      proseRuns.length > 0 &&
-      proseDominantPdfMathSource({
-        text: proseRuns.map((run) => run.text).join(' '),
-        width:
-          Math.max(...proseRuns.map((run) => run.x + run.width)) -
-          Math.min(...proseRuns.map((run) => run.x)),
-        runs: proseRuns,
-      })
-    const exactDetached = extensionRuns.filter((run) =>
-      exactRadicalHosts.has(run),
-    )
-    if (!fallbackProseCue && exactDetached.length === 0) return [preparedLine]
-    const maximumProseFontSize =
-      proseRuns.length > 0
-        ? Math.max(...proseRuns.map((run) => run.fontSize))
-        : 0
-    const proseBaselineCenter =
-      proseRuns.length > 0
-        ? proseRuns.reduce((total, run) => total + run.y + run.height / 2, 0) /
-          proseRuns.length
-        : 0
-    const detachedHosts = new Map<PdfSourceRun, string | null>()
-    const detached = extensionRuns.filter((run) => {
-      const exactHost = exactRadicalHosts.get(run)
-      if (exactHost) {
-        detachedHosts.set(run, exactHost)
-        return true
-      }
-      if (!fallbackProseCue) return false
-      if (
-        run.fontSize > maximumProseFontSize * 0.82 ||
-        Math.abs(run.y + run.height / 2 - proseBaselineCenter) <
-          Math.max(0.002, run.height * 0.3)
-      ) {
-        return false
-      }
-      const candidates = hosts
-        .filter((host) => {
-          if (host === line || host.page !== line.page || !host.id) return false
-          const horizontalGap = Math.max(
-            host.x - (run.x + run.width),
-            run.x - (host.x + host.width),
-            0,
-          )
-          const verticalGap = Math.max(
-            host.y - (run.y + run.height),
-            run.y - (host.y + host.height),
-            0,
-          )
-          return horizontalGap <= 0.02 && verticalGap <= 0.018
-        })
-        .map((host) => ({ host, score: boxDistance(run, host) }))
-        .sort(
-          (left, right) =>
-            left.score - right.score ||
-            sourceLineFlowOrder(left.host, right.host),
-        )
-      const uniquelyHosted =
-        candidates.length > 0 &&
-        (candidates.length === 1 ||
-          candidates[1].score - candidates[0].score > 0.0015)
-      if (uniquelyHosted) {
-        detachedHosts.set(run, candidates[0].host.id!)
-      } else {
-        // A visibly displaced extension glyph is not prose merely because two
-        // nearby formula hosts are tied. Split it into an explicit unresolved
-        // equation obligation; retaining it in the cue would silently publish
-        // the prose while allowing both candidate formulas to omit the glyph.
-        detachedHosts.set(run, null)
-      }
-      return true
-    })
-    if (detached.length === 0) return [preparedLine]
-    const retainedRuns = runs.filter((run) => !detached.includes(run))
-    const fragments: PdfTextLine[] =
-      retainedRuns.length > 0
-        ? [
-            {
-              ...textLineFromRuns(preparedLine, retainedRuns),
-              id: preparedLine.id,
-            },
-          ]
-        : []
-    for (const [index, run] of detached.entries()) {
-      const detachedSourceLineId = sourceLineId(preparedLine, lineIndex)
-      const hostLineId = detachedHosts.get(run)
-      fragments.push({
-        ...textLineFromRuns(preparedLine, [run]),
-        id: `${detachedSourceLineId}-detached-math-${String(index + 1).padStart(3, '0')}-host-${
-          hostLineId ? encodeURIComponent(hostLineId) : 'ambiguous'
-        }`,
-      })
-    }
-    return fragments
-  })
-}
-
-function inlineStackedFragmentParts(line: Pick<PdfTextLine, 'id'>) {
-  const match = line.id?.match(
-    /^(.*-inline-stacked-\d+)-(before|formula|after)$/u,
-  )
-  return match
-    ? {
-        baseId: match[1],
-        part: match[2] as 'before' | 'formula' | 'after',
-      }
-    : null
-}
-
-function inlineStackedFragmentOrder(
-  left: Pick<PdfTextLine, 'id'>,
-  right: Pick<PdfTextLine, 'id'>,
-) {
-  const leftParts = inlineStackedFragmentParts(left)
-  const rightParts = inlineStackedFragmentParts(right)
-  if (!leftParts || !rightParts || leftParts.baseId !== rightParts.baseId) {
-    return null
-  }
-  const order = { before: 0, formula: 1, after: 2 } as const
-  return order[leftParts.part] - order[rightParts.part]
-}
-
-function restoreInlineStackedAtomicUnits(ordered: PdfPageRegion[]) {
-  const units = new Map<
-    string,
-    {
-      page: number
-      invalid: boolean
-      parts: Map<'before' | 'formula' | 'after', PdfPageRegion>
-    }
-  >()
-  const basesByRegionId = new Map<string, Set<string>>()
-
-  for (const region of ordered) {
-    for (const line of region.lines) {
-      const fragment = inlineStackedFragmentParts(line)
-      if (!fragment) continue
-      const unit = units.get(fragment.baseId) ?? {
-        page: region.page,
-        invalid: false,
-        parts: new Map<'before' | 'formula' | 'after', PdfPageRegion>(),
-      }
-      const existing = unit.parts.get(fragment.part)
-      if (unit.page !== region.page || existing !== undefined) {
-        unit.invalid = true
-      }
-      unit.parts.set(fragment.part, region)
-      units.set(fragment.baseId, unit)
-      const regionBases = basesByRegionId.get(region.id) ?? new Set<string>()
-      regionBases.add(fragment.baseId)
-      basesByRegionId.set(region.id, regionBases)
-    }
-  }
-
-  const validUnits = new Map<
-    string,
-    [PdfPageRegion, PdfPageRegion, PdfPageRegion]
-  >()
-  for (const [baseId, unit] of units) {
-    const before = unit.parts.get('before')
-    const formula = unit.parts.get('formula')
-    const after = unit.parts.get('after')
-    const regionIds = new Set(
-      [before, formula, after].flatMap((region) =>
-        region === undefined ? [] : [region.id],
-      ),
-    )
-    const notePartitions = new Set(
-      [before, formula, after].flatMap((region) =>
-        region === undefined
-          ? []
-          : [region.kind === 'footnote' || region.kind === 'endnote'],
-      ),
-    )
-    if (
-      unit.invalid ||
-      !before ||
-      !formula ||
-      !after ||
-      regionIds.size !== 3 ||
-      notePartitions.size !== 1 ||
-      [...regionIds].some(
-        (regionId) => (basesByRegionId.get(regionId)?.size ?? 0) !== 1,
-      )
-    ) {
-      continue
-    }
-    validUnits.set(baseId, [before, formula, after])
-  }
-  if (validUnits.size === 0) return ordered
-
-  const unitByRegionId = new Map<
-    string,
-    [PdfPageRegion, PdfPageRegion, PdfPageRegion]
-  >()
-  for (const unit of validUnits.values()) {
-    for (const region of unit) unitByRegionId.set(region.id, unit)
-  }
-  const emittedRegionIds = new Set<string>()
-  return ordered.flatMap((region) => {
-    if (emittedRegionIds.has(region.id)) return []
-    const unit = unitByRegionId.get(region.id)
-    if (!unit) {
-      emittedRegionIds.add(region.id)
-      return [region]
-    }
-    for (const member of unit) emittedRegionIds.add(member.id)
-    return unit
-  })
-}
-
-function sourceLineFlowOrder(left: PdfTextLine, right: PdfTextLine) {
-  if (left.page !== right.page) return left.page - right.page
-  const inlineOrder = inlineStackedFragmentOrder(left, right)
-  return inlineOrder ?? (left.y - right.y || left.x - right.x)
 }
 
 function emphasizedSourceRun(run: PdfTextLine['runs'][number]) {
@@ -2408,59 +869,6 @@ function noteStratumRunKeys(lines: readonly PdfTextLine[]) {
   return keys
 }
 
-function spread(values: number[]) {
-  return values.length < 2 ? 0 : Math.max(...values) - Math.min(...values)
-}
-
-function distinctLines(values: ClassifiedLine[]): ClassifiedLine[] {
-  return [...new Set(values)]
-}
-
-function alignedBandCount(
-  members: Array<{ left: ClassifiedLine; right: ClassifiedLine }>,
-) {
-  const centers = members
-    .map(
-      ({ left, right }) =>
-        (left.y + left.height / 2 + right.y + right.height / 2) / 2,
-    )
-    .sort((left, right) => left - right)
-  const bands: number[] = []
-  for (const center of centers) {
-    if (bands.every((candidate) => Math.abs(candidate - center) > 0.012)) {
-      bands.push(center)
-    }
-  }
-  return bands.length
-}
-
-function maximumVerticalGap(lines: ClassifiedLine[]) {
-  const ordered = distinctLines(lines).sort(
-    (left, right) => left.y - right.y || left.x - right.x,
-  )
-  if (ordered.length < 2) return Number.POSITIVE_INFINITY
-  return Math.max(
-    ...ordered.slice(1).map((line, index) => {
-      const previous = ordered[index]
-      return line.y - (previous.y + previous.height)
-    }),
-  )
-}
-
-function dominantIndent(lines: ClassifiedLine[]) {
-  const clusters = lines.map((seed) => {
-    const members = lines.filter((line) => Math.abs(line.x - seed.x) <= 0.04)
-    return {
-      members,
-      x: median(members.map((line) => line.x)),
-    }
-  })
-  return clusters.sort(
-    (left, right) =>
-      right.members.length - left.members.length || left.x - right.x,
-  )[0]
-}
-
 function preclassifyMarginNotes(
   lines: ClassifiedLine[],
   pageBodyFontSize: number,
@@ -2520,1347 +928,6 @@ function sourceRaisedFrontMatterAffiliation(line: ClassifiedLine) {
     horizontalGap >= -0.001 &&
     horizontalGap <= 0.012
   )
-}
-
-function endsCaptionSentence(text: string) {
-  return /[.!?][”’'"\])}]*$/u.test(text.trim())
-}
-
-function endsIncompleteWrappedUrl(text: string) {
-  return /(?:https?:\/\/|www\.)\S+\.$/iu.test(text.trim())
-}
-
-function captionLineFontSize(line: PdfTextLine) {
-  const runs = line.runs
-    .filter((run) => run.fontSize > 0 && run.text.trim().length > 0)
-    .map((run) => ({
-      fontSize: run.fontSize,
-      weight: Math.max(1, run.text.replace(/\s/gu, '').length),
-    }))
-    .sort((left, right) => left.fontSize - right.fontSize)
-  if (runs.length === 0) return line.fontSize
-
-  const midpoint = runs.reduce((total, run) => total + run.weight, 0) / 2
-  let cumulative = 0
-  for (const run of runs) {
-    cumulative += run.weight
-    if (cumulative >= midpoint) return run.fontSize
-  }
-  return line.fontSize
-}
-
-export function captionFontFamily(fontName: string) {
-  const normalized = fontName.trim().toLocaleLowerCase()
-  if (/^[a-z][a-z0-9]*_d\d+_f\d+$/u.test(normalized)) {
-    return normalized.replace(/[^a-z0-9]+/gu, '')
-  }
-  return normalized
-    .replace(/^[a-z]{6}\+/iu, '')
-    .replace(/mt$/iu, '')
-    .replace(/ps(?=[-+_,.\s]|$)/iu, '')
-    .replace(
-      /(?:[-+_,.\s]*(?:bold|black|demi(?:bold)?|semibold|medium|regular|roman|book|italic|ital|oblique|obl))+$/iu,
-      '',
-    )
-    .replace(
-      /(?:[-+_,.\s]+(?:reguital|medi|regu|bdit|bdi|bi|bd|it|reg|rm|md|med|lt|sb))+$/iu,
-      '',
-    )
-    .replace(/\d+$/u, '')
-    .replace(/[^a-z0-9]+/gu, '')
-}
-
-function captionLineDominantFontFamily(line: PdfTextLine) {
-  const counts = new Map<string, number>()
-  for (const run of line.runs) {
-    const textLength = run.text.replace(/\s/gu, '').length
-    if (textLength === 0) continue
-    const family = captionFontFamily(run.fontName)
-    if (!family) continue
-    counts.set(family, (counts.get(family) ?? 0) + textLength)
-  }
-  return (
-    [...counts.entries()].sort(
-      ([leftFamily, leftCount], [rightFamily, rightCount]) =>
-        rightCount - leftCount || leftFamily.localeCompare(rightFamily),
-    )[0]?.[0] ?? null
-  )
-}
-
-function captionTypographyCompatible(
-  seed: PdfTextLine,
-  candidate: PdfTextLine,
-) {
-  return (
-    captionLineDominantFontFamily(seed) ===
-    captionLineDominantFontFamily(candidate)
-  )
-}
-
-function captionLaneCompatible(
-  seed: PdfTextLine,
-  previous: PdfTextLine,
-  candidate: PdfTextLine,
-) {
-  const boundary =
-    seed.sourceCaptionLaneBoundary ?? previous.sourceCaptionLaneBoundary
-  const side = seed.sourceCaptionLaneSide ?? previous.sourceCaptionLaneSide
-  if (boundary === undefined || side === undefined) return true
-  if (
-    (candidate.sourceCaptionLaneBoundary !== undefined &&
-      Math.abs(candidate.sourceCaptionLaneBoundary - boundary) > 0.002) ||
-    (candidate.sourceCaptionLaneSide !== undefined &&
-      candidate.sourceCaptionLaneSide !== side)
-  ) {
-    return false
-  }
-  const left = candidate.x
-  const right = candidate.x + candidate.width
-  const tolerance = 0.004
-  if (candidate.sourceCaptionLaneSide === undefined) {
-    if (side === 'left' && right > boundary - tolerance) return false
-    if (side === 'right' && left < boundary + tolerance) return false
-  }
-  if (left < boundary - tolerance && right > boundary + tolerance) {
-    return false
-  }
-  const center = left + candidate.width / 2
-  return side === 'left'
-    ? center <= boundary + tolerance
-    : center >= boundary - tolerance
-}
-
-function completesQuotedCaptionExpression(
-  previous: ClassifiedLine,
-  candidate: ClassifiedLine,
-) {
-  const openQuotes = previous.text.match(/“/gu)?.length ?? 0
-  const closeQuotes = previous.text.match(/”/gu)?.length ?? 0
-  return (
-    openQuotes > closeQuotes &&
-    candidate.text.includes('”') &&
-    endsCaptionSentence(candidate.text)
-  )
-}
-
-function captionContinuationGeometry(
-  seed: ClassifiedLine,
-  previous: ClassifiedLine,
-  candidate: ClassifiedLine,
-  enforceCaptionLane = true,
-) {
-  if (candidate.captionLaneSplitAmbiguous) return false
-  if (
-    (enforceCaptionLane && !captionLaneCompatible(seed, previous, candidate)) ||
-    !captionTypographyCompatible(seed, candidate)
-  ) {
-    return false
-  }
-  const candidateProseWordCount =
-    candidate.text.match(/\p{L}{2,}/gu)?.length ?? 0
-  const equationClassifiedCaptionProse =
-    candidate.kind === 'equation' &&
-    (candidateProseWordCount >= 4 ||
-      (candidateProseWordCount >= 2 && endsCaptionSentence(candidate.text)) ||
-      completesQuotedCaptionExpression(previous, candidate))
-  const footerClassifiedSplitWordContinuation =
-    candidate.kind === 'footer' &&
-    /\p{L}-$|\p{L}‐$/u.test(previous.text.trim()) &&
-    /^\p{L}/u.test(candidate.text.trim())
-  if (
-    candidate.page !== seed.page ||
-    beginsVisualCaption(candidate.text) ||
-    (candidate.kind !== 'body' &&
-      candidate.kind !== 'caption' &&
-      !equationClassifiedCaptionProse &&
-      !footerClassifiedSplitWordContinuation)
-  )
-    return false
-  if (candidate.y <= previous.y + previous.height * 0.35) return false
-
-  const gap = candidate.y - (previous.y + previous.height)
-  const maximumGap = Math.max(
-    0.006,
-    Math.max(previous.height, candidate.height) * 0.72,
-  )
-  if (gap < -0.004 || gap > maximumGap) return false
-
-  const seedFontSize = captionLineFontSize(seed)
-  const candidateFontSize = captionLineFontSize(candidate)
-  const fontRatio =
-    Math.max(seedFontSize, candidateFontSize) /
-    Math.max(1, Math.min(seedFontSize, candidateFontSize))
-  if (fontRatio > 1.12) return false
-
-  const intersection = Math.max(
-    0,
-    Math.min(seed.x + seed.width, candidate.x + candidate.width) -
-      Math.max(seed.x, candidate.x),
-  )
-  const overlapRatio =
-    intersection / Math.max(0.001, Math.min(seed.width, candidate.width))
-  const alignmentTolerance = Math.max(0.025, seed.height * 1.6)
-  const aligned =
-    Math.abs(candidate.x - seed.x) <= alignmentTolerance ||
-    Math.abs(candidate.x + candidate.width - (seed.x + seed.width)) <=
-      alignmentTolerance ||
-    Math.abs(candidate.x + candidate.width / 2 - (seed.x + seed.width / 2)) <=
-      alignmentTolerance
-  return overlapRatio >= 0.72 && aligned
-}
-
-function sourceCaptionLaneSeedPairs(lines: ClassifiedLine[]) {
-  const seeds = lines.filter(
-    (line) =>
-      line.kind === 'caption' &&
-      line.sourceCaptionLaneBoundary !== undefined &&
-      (line.sourceCaptionLaneSide === 'left' ||
-        line.sourceCaptionLaneSide === 'right'),
-  )
-  const candidates = seeds.flatMap((left) => {
-    if (left.sourceCaptionLaneSide !== 'left') return []
-    const rightCandidates = seeds.filter(
-      (right) =>
-        right.sourceCaptionLaneSide === 'right' &&
-        right.page === left.page &&
-        right.sourceCaptionLaneBoundary === left.sourceCaptionLaneBoundary &&
-        Math.abs(left.y + left.height / 2 - (right.y + right.height / 2)) <=
-          Math.max(0.004, left.height * 0.65, right.height * 0.65),
-    )
-    return rightCandidates.length === 1
-      ? [
-          {
-            left,
-            right: rightCandidates[0],
-            boundary: left.sourceCaptionLaneBoundary!,
-          },
-        ]
-      : []
-  })
-  return candidates.filter(
-    ({ left, right, boundary }) =>
-      candidates.filter(
-        (candidate) =>
-          candidate.right.id === right.id && candidate.boundary === boundary,
-      ).length === 1 &&
-      left.x + left.width <= boundary &&
-      right.x >= boundary,
-  )
-}
-
-function splitSourceCaptionLaneContinuations(lines: ClassifiedLine[]) {
-  const output = [...lines]
-  const claimed = new Set<string>()
-  for (const pair of sourceCaptionLaneSeedPairs(output)) {
-    let leftPrevious = pair.left
-    let rightPrevious = pair.right
-    while (true) {
-      const candidates = output.flatMap((line) => {
-        if (
-          claimed.has(line.id) ||
-          line.page !== pair.left.page ||
-          beginsVisualCaption(line.text) ||
-          line.sourceCaptionLaneSide !== undefined ||
-          line.runs.length < 2 ||
-          line.x >= pair.boundary ||
-          line.x + line.width <= pair.boundary
-        ) {
-          return []
-        }
-        const runs = [...line.runs].sort(
-          (left, right) => left.x - right.x || left.y - right.y,
-        )
-        const boundaries = runs.flatMap((rightRun, index) => {
-          if (index === 0) return []
-          const leftRun = runs[index - 1]
-          const leftEdge = leftRun.x + leftRun.width
-          const gap = rightRun.x - leftEdge
-          const center = (leftEdge + rightRun.x) / 2
-          if (
-            gap <
-              Math.max(
-                0.002,
-                Math.min(leftRun.height, rightRun.height) * 0.2,
-              ) ||
-            Math.abs(center - pair.boundary) > 0.02
-          ) {
-            return []
-          }
-          const left = {
-            ...classifiedLineFragment(line, runs.slice(0, index), 'left'),
-            sourceCaptionLaneBoundary: pair.boundary,
-            sourceCaptionLaneSide: 'left' as const,
-          }
-          const right = {
-            ...classifiedLineFragment(line, runs.slice(index), 'right'),
-            sourceCaptionLaneBoundary: pair.boundary,
-            sourceCaptionLaneSide: 'right' as const,
-          }
-          return captionContinuationGeometry(
-            pair.left,
-            leftPrevious,
-            left,
-            false,
-          ) &&
-            captionContinuationGeometry(pair.right, rightPrevious, right, false)
-            ? [{ line, left, right }]
-            : []
-        })
-        if (boundaries.length !== 1) {
-          if (
-            captionContinuationGeometry(pair.left, leftPrevious, line, false) ||
-            captionContinuationGeometry(pair.right, rightPrevious, line, false)
-          ) {
-            line.captionLaneSplitAmbiguous = true
-          }
-          return []
-        }
-        return boundaries
-      })
-      const orderedCandidates = candidates.sort(
-        (left, right) =>
-          Math.max(left.left.y, left.right.y) -
-            Math.max(right.left.y, right.right.y) ||
-          left.line.id.localeCompare(right.line.id),
-      )
-      const first = orderedCandidates[0]
-      if (!first) break
-      const firstCenter = first.line.y + first.line.height / 2
-      const immediateCandidates = orderedCandidates.filter((candidate) => {
-        const center = candidate.line.y + candidate.line.height / 2
-        return (
-          Math.abs(center - firstCenter) <=
-          Math.max(
-            0.004,
-            first.line.height * 0.65,
-            candidate.line.height * 0.65,
-          )
-        )
-      })
-      if (immediateCandidates.length !== 1) {
-        for (const candidate of immediateCandidates) {
-          candidate.line.captionLaneSplitAmbiguous = true
-        }
-        break
-      }
-      const [next] = immediateCandidates
-
-      const index = output.indexOf(next.line)
-      if (index < 0) break
-      output.splice(index, 1, next.left, next.right)
-      claimed.add(next.line.id)
-      leftPrevious = next.left
-      rightPrevious = next.right
-    }
-  }
-  return output
-}
-
-function demoteInlineVisualReferenceContinuations(lines: ClassifiedLine[]) {
-  const ordered = [...lines].sort(
-    (left, right) => left.y - right.y || left.x - right.x,
-  )
-  for (const candidate of ordered.filter(
-    (line) => line.kind === 'caption' && beginsVisualCaption(line.text),
-  )) {
-    const previous = ordered
-      .filter(
-        (line) =>
-          line.id !== candidate.id &&
-          line.page === candidate.page &&
-          line.kind === 'body' &&
-          line.y < candidate.y,
-      )
-      .filter((line) => {
-        const text = line.text.trim()
-        const proseWordCount = text.match(/\p{L}{2,}/gu)?.length ?? 0
-        const unfinishedBodyLine =
-          proseWordCount >= 4 &&
-          /\p{Ll}[\p{L}\p{M}'’]*$/u.test(text) &&
-          !/[.!?;:]\s*$/u.test(text)
-        if (!unfinishedBodyLine) return false
-
-        const gap = candidate.y - (line.y + line.height)
-        const maximumGap = Math.max(
-          0.006,
-          Math.max(line.height, candidate.height) * 0.45,
-        )
-        if (gap < -0.003 || gap > maximumGap) return false
-
-        const previousFontSize = captionLineFontSize(line)
-        const candidateFontSize = captionLineFontSize(candidate)
-        const fontRatio =
-          Math.max(previousFontSize, candidateFontSize) /
-          Math.max(1, Math.min(previousFontSize, candidateFontSize))
-        if (fontRatio > 1.06) return false
-
-        const intersection = Math.max(
-          0,
-          Math.min(line.x + line.width, candidate.x + candidate.width) -
-            Math.max(line.x, candidate.x),
-        )
-        const overlapRatio =
-          intersection / Math.max(0.001, Math.min(line.width, candidate.width))
-        const alignmentTolerance = Math.max(
-          0.014,
-          Math.max(line.height, candidate.height),
-        )
-        return (
-          overlapRatio >= 0.72 &&
-          Math.abs(line.x - candidate.x) <= alignmentTolerance
-        )
-      })
-      .sort(
-        (left, right) =>
-          right.y + right.height - (left.y + left.height) ||
-          Math.abs(left.x - candidate.x) - Math.abs(right.x - candidate.x),
-      )[0]
-    if (!previous) continue
-
-    candidate.kind = 'body'
-    candidate.confidence = Math.min(candidate.confidence, previous.confidence)
-  }
-}
-
-/**
- * PDF text extraction marks only the label-bearing first line of a wrapped
- * caption. Promote its geometrically continuous wrap lines before column
- * classification so a page-spanning table caption does not become a separate
- * `spanning` prose region. The chain stops at whitespace, font, alignment, and
- * paragraph-boundary evidence rather than consuming the following body flow.
- */
-function promoteCaptionContinuations(lines: ClassifiedLine[]) {
-  const claimed = new Set<string>()
-  const seeds = lines
-    .filter((line) => line.kind === 'caption')
-    .sort((left, right) => left.y - right.y || left.x - right.x)
-
-  for (const seed of seeds) {
-    let previous = seed
-    const alignmentTolerance = Math.max(0.025, seed.height * 1.6)
-    const seedFontSize = captionLineFontSize(seed)
-    let widestLine = Math.max(
-      seed.width,
-      ...lines
-        .filter((line) => {
-          const lineFontSize = captionLineFontSize(line)
-          const fontRatio =
-            Math.max(seedFontSize, lineFontSize) /
-            Math.max(1, Math.min(seedFontSize, lineFontSize))
-          return (
-            line.page === seed.page &&
-            fontRatio <= 1.12 &&
-            Math.abs(line.x - seed.x) <= alignmentTolerance
-          )
-        })
-        .map((line) => line.width),
-    )
-    while (true) {
-      const previousIsShortConclusion =
-        endsCaptionSentence(previous.text) && previous.width < widestLine * 0.82
-      if (previousIsShortConclusion) break
-
-      const candidate = lines
-        .filter(
-          (line) =>
-            !claimed.has(line.id) &&
-            captionContinuationGeometry(seed, previous, line),
-        )
-        .sort(
-          (left, right) =>
-            left.y - right.y ||
-            Math.abs(left.x - seed.x) - Math.abs(right.x - seed.x),
-        )[0]
-      if (!candidate) break
-
-      const paragraphIndent = candidate.x - seed.x
-      if (
-        endsCaptionSentence(previous.text) &&
-        paragraphIndent > Math.max(0.018, seed.height * 1.25)
-      ) {
-        break
-      }
-
-      const fragmentParts = inlineStackedFragmentParts(candidate)
-      const fragmentSiblings = fragmentParts
-        ? lines
-            .filter(
-              (line) =>
-                inlineStackedFragmentParts(line)?.baseId ===
-                fragmentParts.baseId,
-            )
-            .sort(sourceLineFlowOrder)
-        : [candidate]
-      const continuationLines = fragmentSiblings.every(
-        (line) => !claimed.has(line.id),
-      )
-        ? fragmentSiblings
-        : [candidate]
-      for (const continuation of continuationLines) {
-        const continuationFragment = inlineStackedFragmentParts(continuation)
-        if (continuationFragment?.part === 'formula') {
-          // Keep the exact source geometry in the caption's reading-order
-          // neighborhood, but do not relabel a provably two-dimensional math
-          // fragment as linear caption prose. The visual pass owns this line
-          // as an unresolved equation crop obligation.
-          continuation.kind = 'equation'
-          delete continuation.captionContinuationSeedId
-        } else {
-          continuation.kind = 'caption'
-          continuation.captionContinuationSeedId = seed.id
-        }
-        continuation.confidence = Math.min(
-          seed.confidence,
-          continuation.confidence,
-        )
-        claimed.add(continuation.id)
-        widestLine = Math.max(widestLine, continuation.width)
-      }
-      previous = continuationLines.at(-1)!
-    }
-  }
-}
-
-function horizontalOverlapRatio(
-  left: Pick<PdfTextLine, 'x' | 'width'>,
-  right: Pick<PdfTextLine, 'x' | 'width'>,
-) {
-  const overlap = Math.max(
-    0,
-    Math.min(left.x + left.width, right.x + right.width) -
-      Math.max(left.x, right.x),
-  )
-  return overlap / Math.max(0.001, Math.min(left.width, right.width))
-}
-
-/**
- * Side-by-side panels are emitted row-first by PDF.js: panel A line 1,
- * panel B line 1, panel A line 2, panel B line 2. Without the native-panel
- * and shared-caption proof below, sequential region joining assigns each
- * continuation to the opposite panel. Preserve the source lane before region
- * grouping so the visual pass can own complete `(a)` / `(b)` labels.
- */
-function markAdjacentPanelLabelContinuations(
-  lines: ClassifiedLine[],
-  objects: PdfPageAnalysis['objects'],
-) {
-  const images = (objects ?? []).filter(
-    (object) => object.kind === 'image' && object.role !== 'scan-source',
-  )
-  if (images.length < 2) return
-
-  for (const caption of lines.filter(
-    (line) => line.kind === 'caption' && beginsVisualCaption(line.text),
-  )) {
-    const eligibleImages = images.filter(
-      (image) =>
-        image.box.y + image.box.height <= caption.y + 0.004 &&
-        caption.y - (image.box.y + image.box.height) <= 0.12,
-    )
-    if (eligibleImages.length < 2) continue
-    const closestBottom = Math.max(
-      ...eligibleImages.map((image) => image.box.y + image.box.height),
-    )
-    const bandReference = eligibleImages.reduce((closest, candidate) =>
-      candidate.box.y + candidate.box.height >
-      closest.box.y + closest.box.height
-        ? candidate
-        : closest,
-    )
-    const panels = eligibleImages
-      .filter((image) => {
-        const bottom = image.box.y + image.box.height
-        const verticalOverlap = Math.max(
-          0,
-          Math.min(
-            image.box.y + image.box.height,
-            bandReference.box.y + bandReference.box.height,
-          ) - Math.max(image.box.y, bandReference.box.y),
-        )
-        return (
-          closestBottom - bottom <= 0.02 &&
-          verticalOverlap >=
-            Math.min(image.box.height, bandReference.box.height) * 0.7
-        )
-      })
-      .sort((left, right) => left.box.x - right.box.x)
-    if (panels.length < 2) continue
-    const panelLeft = Math.min(...panels.map((panel) => panel.box.x))
-    const panelRight = Math.max(
-      ...panels.map((panel) => panel.box.x + panel.box.width),
-    )
-    if (
-      panelRight - panelLeft < 0.5 ||
-      caption.width < (panelRight - panelLeft) * 0.75
-    ) {
-      continue
-    }
-
-    const panelBottom = Math.max(
-      ...panels.map((panel) => panel.box.y + panel.box.height),
-    )
-    const seeds = lines
-      .filter(
-        (line) =>
-          line.kind === 'body' &&
-          /^\s*\([a-z0-9ivxlcdm]+\)\s+\p{L}/iu.test(line.text) &&
-          line.y >= panelBottom - 0.012 &&
-          line.y + line.height <= caption.y + 0.004,
-      )
-      .sort((left, right) => left.x - right.x)
-    if (seeds.length !== panels.length) continue
-    const seedPanels = seeds.map((seed) => {
-      const overlaps = panels.map((panel) =>
-        horizontalOverlapRatio(seed, panel.box),
-      )
-      const best = Math.max(...overlaps)
-      const bestIndex = overlaps.indexOf(best)
-      const runnerUp = Math.max(
-        ...overlaps.filter((_overlap, index) => index !== bestIndex),
-        0,
-      )
-      return best >= 0.7 && best - runnerUp >= 0.2 ? bestIndex : -1
-    })
-    if (
-      seedPanels.some((index) => index < 0) ||
-      new Set(seedPanels).size !== panels.length
-    ) {
-      continue
-    }
-
-    const claimed = new Set(seeds.map((seed) => seed.id))
-    for (const [seedIndex, seed] of seeds.entries()) {
-      const panel = panels[seedPanels[seedIndex]]
-      let previous = seed
-      while (true) {
-        const candidates = lines
-          .filter((candidate) => {
-            if (
-              claimed.has(candidate.id) ||
-              candidate.kind !== 'body' ||
-              candidate.page !== seed.page ||
-              candidate.y < previous.y + previous.height - 0.002 ||
-              candidate.y + candidate.height > caption.y + 0.004 ||
-              /^\s*\([a-z0-9ivxlcdm]+\)\s+\p{L}/iu.test(candidate.text)
-            ) {
-              return false
-            }
-            const gap = candidate.y - (previous.y + previous.height)
-            const fontRatio =
-              Math.max(seed.fontSize, candidate.fontSize) /
-              Math.max(1, Math.min(seed.fontSize, candidate.fontSize))
-            return (
-              gap <= Math.max(0.014, seed.height * 1.25) &&
-              fontRatio <= 1.18 &&
-              horizontalOverlapRatio(candidate, panel.box) >= 0.7 &&
-              panels.filter(
-                (otherPanel) =>
-                  horizontalOverlapRatio(candidate, otherPanel.box) >= 0.7,
-              ).length === 1
-            )
-          })
-          .sort(
-            (left, right) =>
-              left.y - right.y ||
-              Math.abs(left.x - panel.box.x) - Math.abs(right.x - panel.box.x),
-          )
-        if (candidates.length === 0) break
-        const candidate = candidates[0]
-        candidate.panelLabelContinuationSeedId = seed.id
-        candidate.confidence = Math.min(seed.confidence, candidate.confidence)
-        claimed.add(candidate.id)
-        previous = candidate
-      }
-    }
-  }
-}
-
-function promoteNoteContinuations(lines: ClassifiedLine[]) {
-  const claimed = new Set<string>()
-  const seeds = lines
-    .filter(
-      (line) =>
-        (line.kind === 'footnote' || line.kind === 'endnote') && line.y >= 0.65,
-    )
-    .sort((left, right) => left.y - right.y || left.x - right.x)
-
-  for (const seed of seeds) {
-    let previous = seed
-    while (
-      !endsCaptionSentence(previous.text) ||
-      endsIncompleteWrappedUrl(previous.text)
-    ) {
-      const candidate = lines
-        .filter((line) => {
-          if (
-            line.page !== seed.page ||
-            !(
-              line.kind === 'body' ||
-              line.kind === 'footer' ||
-              ((line.kind === 'footnote' || line.kind === 'endnote') &&
-                line.noteLabel === null)
-            ) ||
-            claimed.has(line.id) ||
-            line.y <= previous.y
-          ) {
-            return false
-          }
-          const verticalGap = line.y - (previous.y + previous.height)
-          const fontTolerance = Math.max(0.6, seed.fontSize * 0.08)
-          return (
-            verticalGap <= Math.max(0.008, seed.height * 0.8) &&
-            Math.abs(line.x - seed.x) <= 0.04 &&
-            Math.abs(line.fontSize - seed.fontSize) <= fontTolerance
-          )
-        })
-        .sort(
-          (left, right) =>
-            left.y - right.y ||
-            Math.abs(left.x - seed.x) - Math.abs(right.x - seed.x),
-        )[0]
-      if (!candidate) break
-      candidate.kind = seed.kind
-      candidate.noteContinuationSeedId = seed.id
-      candidate.confidence = Math.min(seed.confidence, candidate.confidence)
-      claimed.add(candidate.id)
-      previous = candidate
-    }
-  }
-}
-
-type TabularGridFragment = {
-  left: number
-  right: number
-  text: string
-}
-
-type TabularGridBand = {
-  center: number
-  lines: ClassifiedLine[]
-  fragments: TabularGridFragment[]
-  column: Extract<PdfRegionColumn, 'left' | 'right' | 'span'>
-}
-
-function tabularGridFragments(lines: ClassifiedLine[]) {
-  const fragments: TabularGridFragment[] = []
-  const runs = lines
-    .flatMap((line) => line.runs.filter((run) => run.text.trim()))
-    .sort((left, right) => left.x - right.x)
-  for (const run of runs) {
-    const right = run.x + run.width
-    const previous = fragments.at(-1)
-    const joinsPrevious =
-      previous && run.x <= previous.right + Math.max(0.006, run.height * 0.5)
-    if (previous && joinsPrevious) {
-      previous.right = Math.max(previous.right, right)
-      previous.text = `${previous.text} ${run.text}`
-      continue
-    }
-    fragments.push({
-      left: run.x,
-      right,
-      text: run.text,
-    })
-  }
-  return fragments
-}
-
-function hasParallelProseColumnFragments(
-  fragments: readonly TabularGridFragment[],
-) {
-  const left = fragments.filter((fragment) => fragment.right <= 0.49)
-  const right = fragments.filter((fragment) => fragment.left >= 0.51)
-  if (left.length === 0 || right.length === 0) return false
-  const leftCoverage = left.at(-1)!.right - left[0].left
-  const rightCoverage = right.at(-1)!.right - right[0].left
-  const gutter = right[0].left - left.at(-1)!.right
-  const wordCount = (side: readonly TabularGridFragment[]) =>
-    side.reduce(
-      (total, fragment) =>
-        total + (fragment.text.match(/[\p{L}\p{N}]+/gu)?.length ?? 0),
-      0,
-    )
-  return (
-    leftCoverage >= 0.28 &&
-    rightCoverage >= 0.28 &&
-    gutter >= 0.015 &&
-    wordCount(left) >= 4 &&
-    wordCount(right) >= 4
-  )
-}
-
-function hasParallelProseColumnLines(lines: readonly ClassifiedLine[]) {
-  const proseLine = (line: ClassifiedLine) =>
-    line.width >= 0.18 && (line.text.match(/[\p{L}\p{N}]+/gu)?.length ?? 0) >= 3
-  const left = lines.some(
-    (line) => line.x < 0.49 && line.x + line.width <= 0.5 && proseLine(line),
-  )
-  const right = lines.some(
-    (line) => line.x >= 0.5 && line.x + line.width > 0.51 && proseLine(line),
-  )
-  return left && right
-}
-
-function strictTabularNumericCell(text: string) {
-  return /^(?:[<>≤≥~≈])?[+\-−]?(?:\d+(?:[.,]\d+)?|\.\d+)(?:(?:[eE][+\-−]?\d+)|(?:[x×]10[+\-−]?\d+))?(?:%|[x×])?$/u.test(
-    text.replace(/\s+/gu, ''),
-  )
-}
-
-function denseTabularGridBand(band: Omit<TabularGridBand, 'fragments'>) {
-  const fragments = tabularGridFragments(band.lines)
-  if (fragments.length < 4) return null
-  const coverage = fragments.at(-1)!.right - fragments[0].left
-  const compactFragments = fragments.filter((fragment) => {
-    const wordCount = fragment.text.match(/[\p{L}\p{N}]+/gu)?.length ?? 0
-    return fragment.right - fragment.left <= 0.18 || wordCount <= 3
-  }).length
-  const minimumCoverage = band.column === 'span' ? 0.45 : 0.2
-  const coversColumnLane =
-    band.column === 'span' ||
-    (band.column === 'left'
-      ? fragments.at(-1)!.right >= 0.42
-      : fragments[0].left <= 0.58)
-  const provesLaneNumericMatrixBody =
-    band.column === 'span' ||
-    (!strictTabularNumericCell(fragments[0].text) &&
-      /\p{L}/u.test(fragments[0].text) &&
-      fragments
-        .slice(1)
-        .every((fragment) => strictTabularNumericCell(fragment.text)))
-  if (
-    coverage < minimumCoverage ||
-    compactFragments < fragments.length * 0.75 ||
-    !coversColumnLane ||
-    !provesLaneNumericMatrixBody
-  ) {
-    return null
-  }
-  return { ...band, fragments }
-}
-
-/**
- * Dense table rows can repeat many source-run anchors while presenting only
- * two or three grouped text lines. Those repeated cell gaps are not evidence
- * for page columns. Identify compact full-width grids, or numeric grids wholly
- * inside one page lane, with at least four recurring anchors in three adjacent
- * row bands; ordinary two-column prose therefore remains eligible to prove its
- * real gutter.
- */
-function denseTabularGridBands(lines: ClassifiedLine[]) {
-  const bands: Array<Omit<TabularGridBand, 'fragments' | 'column'>> = []
-  for (const line of lines
-    .filter((candidate) => candidate.kind === 'body' && candidate.text.trim())
-    .sort((left, right) => left.y - right.y || left.x - right.x)) {
-    const center = line.y + line.height / 2
-    const band = bands.find(
-      (candidate) =>
-        Math.abs(candidate.center - center) <=
-        Math.max(0.007, line.height * 0.7),
-    )
-    if (band) {
-      band.lines.push(line)
-      band.center = median(
-        band.lines.map((candidate) => candidate.y + candidate.height / 2),
-      )
-    } else {
-      bands.push({ center, lines: [line] })
-    }
-  }
-
-  const denseBands = bands.flatMap<TabularGridBand>((band) => {
-    const spanning = denseTabularGridBand({ ...band, column: 'span' })
-    if (
-      spanning &&
-      !hasParallelProseColumnLines(spanning.lines) &&
-      !hasParallelProseColumnFragments(spanning.fragments)
-    ) {
-      return [spanning]
-    }
-    // A table can occupy one page column while prose continues in the other.
-    // Keep only whole PDF.js-backed lines on each side of the page center;
-    // never split a source run or synthesize a character/cell box.
-    return [
-      {
-        column: 'left' as const,
-        lines: band.lines.filter((line) => line.x + line.width <= 0.49),
-      },
-      {
-        column: 'right' as const,
-        lines: band.lines.filter((line) => line.x >= 0.51),
-      },
-    ]
-      .filter((candidate) => candidate.lines.length > 0)
-      .flatMap((candidate) => {
-        const dense = denseTabularGridBand({
-          center: band.center,
-          lines: candidate.lines,
-          column: candidate.column,
-        })
-        return dense ? [dense] : []
-      })
-  })
-
-  const recurringBands = denseBands.filter((band) => {
-    const recurringFragments = band.fragments.filter((fragment) => {
-      const matchingBands = denseBands.filter(
-        (candidate) =>
-          candidate !== band &&
-          candidate.column === band.column &&
-          candidate.fragments.some(
-            (candidateFragment) =>
-              Math.abs(candidateFragment.left - fragment.left) <= 0.025 ||
-              Math.abs(candidateFragment.right - fragment.right) <= 0.025,
-          ),
-      )
-      return matchingBands.length >= 2
-    })
-    return recurringFragments.length >= 4
-  })
-
-  const adjacentClusters: TabularGridBand[][] = []
-  for (const band of recurringBands.sort(
-    (left, right) =>
-      left.column.localeCompare(right.column) || left.center - right.center,
-  )) {
-    const cluster = adjacentClusters.at(-1)
-    if (
-      cluster &&
-      cluster[0].column === band.column &&
-      band.center - cluster.at(-1)!.center <= 0.08
-    ) {
-      cluster.push(band)
-    } else {
-      adjacentClusters.push([band])
-    }
-  }
-
-  const entries = adjacentClusters.flatMap((cluster, clusterIndex) =>
-    cluster.length < 3
-      ? []
-      : cluster.flatMap((band, bandIndex) =>
-          band.lines.map(
-            (line) =>
-              ({
-                lineId: line.id,
-                bandId: `tabular-grid-${cluster[0].column}-${clusterIndex + 1}-band-${bandIndex + 1}`,
-                column: cluster[0].column,
-              }) as const,
-          ),
-        ),
-  )
-  return {
-    bands: new Map(entries.map((entry) => [entry.lineId, entry.bandId])),
-    columns: new Map(
-      entries.map((entry) => [entry.lineId, entry.column] as const),
-    ),
-  }
-}
-
-function detectColumns(lines: ClassifiedLine[]): ColumnLayout {
-  const sideLines = lines.filter((line) => line.kind === 'side')
-  const tabularGrid = denseTabularGridBands(lines)
-  const tabularGridBands = tabularGrid.bands
-  const tabularGridColumns = tabularGrid.columns
-  const candidates = lines.filter(
-    (line) =>
-      line.kind === 'body' &&
-      line.text.length > 1 &&
-      line.width < 0.72 &&
-      !tabularGridBands.has(line.id),
-  )
-  const gaps: Array<{
-    split: number
-    left: ClassifiedLine
-    right: ClassifiedLine
-  }> = []
-  for (let index = 0; index < candidates.length; index += 1) {
-    for (
-      let nextIndex = index + 1;
-      nextIndex < candidates.length;
-      nextIndex += 1
-    ) {
-      const first = candidates[index]
-      const second = candidates[nextIndex]
-      const left = first.x <= second.x ? first : second
-      const right = left === first ? second : first
-      const centerDifference = Math.abs(
-        left.y + left.height / 2 - (right.y + right.height / 2),
-      )
-      const gap = right.x - (left.x + left.width)
-      if (
-        centerDifference > Math.max(0.007, left.height, right.height) ||
-        gap <= Math.max(0.01, Math.min(left.height, right.height) * 0.6)
-      ) {
-        continue
-      }
-      gaps.push({ split: left.x + left.width + gap / 2, left, right })
-    }
-  }
-  if (gaps.length === 0) {
-    const bodyIndent = spread(candidates.map((line) => line.x))
-    if (sideLines.length > 0 && candidates.length >= 2 && bodyIndent <= 0.04) {
-      const bodyMedian = median(candidates.map((line) => line.fontSize))
-      const sideMedian = median(sideLines.map((line) => line.fontSize))
-      return {
-        split: null,
-        accepted: false,
-        ambiguous: false,
-        tabularGridBands,
-        tabularGridColumns,
-        resolution: {
-          policyVersion: READING_ORDER_RESOLUTION_POLICY_VERSION,
-          ambiguityClass: 'single-column-with-margin-notes',
-          status: 'resolved',
-          confidence: 0.94,
-          threshold: READING_ORDER_RESOLUTION_THRESHOLD,
-          evidence: [
-            {
-              code: 'font-metrics',
-              detail: `Margin median font ${rounded(sideMedian)} is subordinate to body median font ${rounded(bodyMedian)}.`,
-            },
-            {
-              code: 'indentation-continuity',
-              detail: `The retained body flow has normalized indentation spread ${rounded(bodyIndent)}.`,
-            },
-            {
-              code: 'margin-note-exclusion',
-              detail: `${sideLines.length} aligned narrow side region${sideLines.length === 1 ? '' : 's'} remain outside canonical reading order.`,
-            },
-          ],
-        },
-      }
-    }
-    return {
-      split: null,
-      accepted: false,
-      ambiguous: false,
-      tabularGridBands,
-      tabularGridColumns,
-      resolution: null,
-    }
-  }
-
-  const clusters = gaps.map((seed) => {
-    const members = gaps.filter(
-      (candidate) => Math.abs(candidate.split - seed.split) <= 0.05,
-    )
-    return {
-      members,
-      split: median(members.map((member) => member.split)),
-    }
-  })
-  const rankedClusters = clusters.sort(
-    (left, right) =>
-      right.members.length - left.members.length || left.split - right.split,
-  )
-  let strongest = rankedClusters[0]
-  const crossingCandidateCount = (split: number) =>
-    candidates.filter((line) => {
-      const tolerance = Math.max(0.006, Math.min(0.018, line.height * 0.4))
-      return (
-        line.x < split - tolerance && line.x + line.width > split + tolerance
-      )
-    }).length
-  const strongestCrossingCount = crossingCandidateCount(strongest.split)
-  // A compact grid can repeat an internal cell gap more often than the real
-  // page gutter. Prefer only another well-supported central gap, and only
-  // when it partitions more source lines instead of cutting through them.
-  const centralProseGutter = rankedClusters
-    .filter(
-      (cluster) =>
-        cluster.split >= 0.4 &&
-        cluster.split <= 0.6 &&
-        alignedBandCount(cluster.members) >= 2 &&
-        cluster.members.length >=
-          Math.max(2, Math.ceil(strongest.members.length * 0.4)),
-    )
-    .map((cluster) => ({
-      ...cluster,
-      crossingCount: crossingCandidateCount(cluster.split),
-    }))
-    .sort(
-      (left, right) =>
-        left.crossingCount - right.crossingCount ||
-        right.members.length - left.members.length ||
-        Math.abs(left.split - 0.5) - Math.abs(right.split - 0.5),
-    )[0]
-  if (
-    centralProseGutter &&
-    (strongest.split < 0.4 ||
-      strongest.split > 0.6 ||
-      centralProseGutter.crossingCount < strongestCrossingCount)
-  ) {
-    strongest = centralProseGutter
-  }
-  const selectedCrossingCount = crossingCandidateCount(strongest.split)
-  const tolerance = 0.008
-  const leftLines = candidates.filter(
-    (line) => line.x + line.width <= strongest.split + tolerance,
-  )
-  const rightLines = candidates.filter(
-    (line) => line.x >= strongest.split - tolerance,
-  )
-  const leftRange = [
-    Math.min(...leftLines.map((line) => line.y)),
-    Math.max(...leftLines.map((line) => line.y + line.height)),
-  ]
-  const rightRange = [
-    Math.min(...rightLines.map((line) => line.y)),
-    Math.max(...rightLines.map((line) => line.y + line.height)),
-  ]
-  const overlapsVertically =
-    leftLines.length > 0 &&
-    rightLines.length > 0 &&
-    Math.min(leftRange[1], rightRange[1]) >
-      Math.max(leftRange[0], rightRange[0])
-  const pairedLeft = distinctLines(
-    strongest.members.map((member) => member.left),
-  )
-  const pairedRight = distinctLines(
-    strongest.members.map((member) => member.right),
-  )
-  const alignedBands = alignedBandCount(strongest.members)
-  if (
-    overlapsVertically &&
-    alignedBands < 2 &&
-    pairedLeft.length >= 2 &&
-    pairedRight.length >= 2
-  ) {
-    return {
-      split: null,
-      accepted: false,
-      ambiguous: false,
-      tabularGridBands,
-      tabularGridColumns,
-      resolution: {
-        policyVersion: READING_ORDER_RESOLUTION_POLICY_VERSION,
-        ambiguityClass: 'fragmented-inline-cluster',
-        status: 'resolved',
-        confidence: 0.92,
-        threshold: READING_ORDER_RESOLUTION_THRESHOLD,
-        evidence: [
-          {
-            code: 'block-adjacency',
-            detail:
-              'Separated fragments occupy one horizontal band and do not establish adjacent vertical column blocks.',
-          },
-        ],
-      },
-    }
-  }
-
-  const splitDeviation = Math.max(
-    ...strongest.members.map((member) =>
-      Math.abs(member.split - strongest.split),
-    ),
-  )
-  const minimumGutter = Math.min(
-    ...strongest.members.map(
-      (member) => member.right.x - (member.left.x + member.left.width),
-    ),
-  )
-  const gutterStable =
-    alignedBands >= 2 &&
-    strongest.split >= 0.25 &&
-    strongest.split <= 0.75 &&
-    splitDeviation <= 0.025 &&
-    minimumGutter >= 0.01
-  const leftIndentSpread = spread(pairedLeft.map((line) => line.x))
-  const rightIndentSpread = spread(pairedRight.map((line) => line.x))
-  const leftWidthSpread = spread(pairedLeft.map((line) => line.width))
-  const rightWidthSpread = spread(pairedRight.map((line) => line.width))
-  const indentationContinuous =
-    leftIndentSpread <= 0.035 &&
-    rightIndentSpread <= 0.035 &&
-    leftWidthSpread <= 0.08 &&
-    rightWidthSpread <= 0.08
-  const leftFont = median(pairedLeft.map((line) => line.fontSize))
-  const rightFont = median(pairedRight.map((line) => line.fontSize))
-  const fontRatio =
-    Math.max(leftFont, rightFont) / Math.max(1, Math.min(leftFont, rightFont))
-  const fontCompatible = fontRatio <= 1.15
-  const leftGap = maximumVerticalGap(pairedLeft)
-  const rightGap = maximumVerticalGap(pairedRight)
-  const blocksAdjacent =
-    alignedBands >= 2 && leftGap <= 0.12 && rightGap <= 0.12
-
-  const crossesSplit = (line: ClassifiedLine) =>
-    line.x < strongest.split - 0.04 &&
-    line.x + line.width > strongest.split + 0.04
-  const captions = lines.filter(
-    (line) => line.kind === 'caption' && crossesSplit(line),
-  )
-  const spanning = lines.filter(
-    (line) => line.kind === 'body' && crossesSplit(line),
-  )
-  const footnotes = lines.filter(
-    (line) => line.kind === 'footnote' || line.kind === 'endnote',
-  )
-  const ambiguityClass: PdfReadingOrderAmbiguityClass = lines.some((line) =>
-    /^(?:references|bibliography)$/i.test(line.text.trim()),
-  )
-    ? 'dense-reference-section'
-    : footnotes.length > 0
-      ? 'footnote-band'
-      : captions.length > 0
-        ? 'two-column-with-spanning-float'
-        : spanning.length > 0
-          ? 'mixed-single-two-column'
-          : 'sparse-column-gutter'
-  const evidence: PdfReadingOrderEvidence[] = []
-  let confidence = 0.45
-  if (gutterStable) {
-    confidence += 0.2
-    evidence.push({
-      code: 'column-gutter',
-      detail: `${alignedBands} aligned bands repeat a normalized gutter at ${rounded(strongest.split)} with maximum deviation ${rounded(splitDeviation)}; ${selectedCrossingCount} candidate line${selectedCrossingCount === 1 ? '' : 's'} cross the selected split.`,
-    })
-  }
-  if (fontCompatible) {
-    confidence += 0.1
-    evidence.push({
-      code: 'font-metrics',
-      detail: `Left and right median font sizes are ${rounded(leftFont)} and ${rounded(rightFont)} (ratio ${rounded(fontRatio)}).`,
-    })
-  }
-  if (indentationContinuous) {
-    confidence += 0.1
-    evidence.push({
-      code: 'indentation-continuity',
-      detail: `Column indentation spreads are ${rounded(leftIndentSpread)} left and ${rounded(rightIndentSpread)} right; width spreads are ${rounded(leftWidthSpread)} and ${rounded(rightWidthSpread)}.`,
-    })
-  }
-  if (blocksAdjacent) {
-    confidence += 0.1
-    evidence.push({
-      code: 'block-adjacency',
-      detail: `Maximum normalized within-column gaps are ${rounded(leftGap)} left and ${rounded(rightGap)} right.`,
-    })
-  }
-  if (captions.length > 0) {
-    confidence += 0.03
-    evidence.push({
-      code: 'caption-proximity',
-      detail: `${captions.length} caption region${captions.length === 1 ? '' : 's'} cross the stable gutter as a page-spanning boundary.`,
-    })
-  }
-  if (spanning.length > 0) {
-    confidence += 0.02
-    evidence.push({
-      code: 'spanning-boundary',
-      detail: `${spanning.length} body region${spanning.length === 1 ? '' : 's'} cross the stable gutter and delimit column bands.`,
-    })
-  }
-  if (footnotes.length > 0) {
-    confidence += 0.02
-    evidence.push({
-      code: 'footnote-band',
-      detail: `${footnotes.length} smaller-font note region${footnotes.length === 1 ? '' : 's'} form a separated lower band after body flow.`,
-    })
-  }
-  confidence = rounded(Math.min(confidence, 0.99))
-  const coreEvidenceCount = [
-    fontCompatible,
-    indentationContinuous,
-    blocksAdjacent,
-  ].filter(Boolean).length
-  const repeatedGeometry =
-    overlapsVertically &&
-    alignedBands >= 3 &&
-    pairedLeft.length >= 3 &&
-    pairedRight.length >= 3
-  const resolvedSparseGeometry =
-    overlapsVertically &&
-    gutterStable &&
-    coreEvidenceCount >= 2 &&
-    confidence >= READING_ORDER_RESOLUTION_THRESHOLD
-  const accepted = repeatedGeometry || resolvedSparseGeometry
-  const ambiguous =
-    !accepted &&
-    overlapsVertically &&
-    alignedBands >= 2 &&
-    pairedLeft.length >= 2 &&
-    pairedRight.length >= 2
-  return {
-    split: accepted || ambiguous ? rounded(strongest.split) : null,
-    accepted,
-    ambiguous,
-    tabularGridBands,
-    tabularGridColumns,
-    resolution:
-      resolvedSparseGeometry || ambiguous
-        ? {
-            policyVersion: READING_ORDER_RESOLUTION_POLICY_VERSION,
-            ambiguityClass,
-            status: resolvedSparseGeometry ? 'resolved' : 'ambiguous',
-            confidence,
-            threshold: READING_ORDER_RESOLUTION_THRESHOLD,
-            evidence,
-          }
-        : null,
-  }
-}
-
-function columnFor(
-  line: PdfTextLine & { id?: string },
-  layout: ColumnLayout,
-): PdfRegionColumn {
-  const tabularColumn = line.id
-    ? layout.tabularGridColumns?.get(line.id)
-    : undefined
-  if (tabularColumn && tabularColumn !== 'span') return tabularColumn
-  if (layout.split === null) return 'single'
-  if (tabularColumn) return tabularColumn
-  const tolerance = Math.max(0.006, Math.min(0.018, line.height * 0.4))
-  if (line.x + line.width <= layout.split + tolerance) return 'left'
-  if (line.x >= layout.split - tolerance) return 'right'
-  const hangsIntoGutterFromRight =
-    line.x >= layout.split - Math.max(0.02, tolerance * 2) &&
-    line.width <= 0.48 &&
-    line.x + line.width >= layout.split + Math.max(0.12, line.width * 0.35) &&
-    line.x + line.width <= 0.96
-  if (hangsIntoGutterFromRight) return 'right'
-  return 'span'
-}
-
-function classifiedLineFragment(
-  source: ClassifiedLine,
-  runs: ClassifiedLine['runs'],
-  suffix: 'left' | 'right',
-): ClassifiedLine {
-  const orderedRuns = [...runs].sort(
-    (left, right) => left.x - right.x || left.y - right.y,
-  )
-  const left = Math.min(...orderedRuns.map((run) => run.x))
-  const top = Math.min(...orderedRuns.map((run) => run.y))
-  const right = Math.max(...orderedRuns.map((run) => run.x + run.width))
-  const bottom = Math.max(...orderedRuns.map((run) => run.y + run.height))
-  const sourceSequenceIndexes = orderedRuns.flatMap((run) =>
-    run.sourceSequenceIndex === undefined ? [] : [run.sourceSequenceIndex],
-  )
-  const sourceFragmentLineage =
-    sourceSequenceIndexes.length === orderedRuns.length &&
-    new Set(sourceSequenceIndexes).size === sourceSequenceIndexes.length
-      ? {
-          algorithm: 'source-run-fragment-v1' as const,
-          sourceLineId: source.id,
-          fragment: `cross-gutter-${suffix}` as const,
-          sourceSequenceIndexes,
-        }
-      : undefined
-  return {
-    ...source,
-    id: `${source.id}-${suffix}`,
-    text: mergePdfRunText(orderedRuns),
-    x: left,
-    y: top,
-    width: right - left,
-    height: bottom - top,
-    fontSize: Math.max(...orderedRuns.map((run) => run.fontSize)),
-    runs: orderedRuns,
-    ...(sourceFragmentLineage ? { sourceFragmentLineage } : {}),
-  }
 }
 
 export function splitRunBackedCrossGutterProse(
@@ -4052,21 +1119,6 @@ function splitRunBackedSymbolicNoteDefinitions(region: PdfPageRegion) {
       lines: [fragmentLine],
     }
   })
-}
-
-function splitNoteDefinitionOrder(left: PdfPageRegion, right: PdfPageRegion) {
-  const parts = (id: string) => {
-    const match = id.match(/^(.*)-note-(\d{3})$/u)
-    return {
-      base: match?.[1] ?? id,
-      ordinal: match ? Number(match[2]) : 1,
-    }
-  }
-  const leftParts = parts(left.id)
-  const rightParts = parts(right.id)
-  return leftParts.base === rightParts.base
-    ? leftParts.ordinal - rightParts.ordinal
-    : null
 }
 
 function standaloneSectionHeading(line: ClassifiedLine) {
@@ -5532,304 +2584,6 @@ function makeObjectRegions(
         }
       }),
   )
-}
-
-function columnOrdered(regions: PdfPageRegion[]) {
-  const readingAnchor = (region: PdfPageRegion) => {
-    const firstSourceLine = [...region.lines]
-      .filter((line) => line.text.trim())
-      .sort(
-        (left, right) => left.box.y - right.box.y || left.box.x - right.box.x,
-      )[0]
-    return firstSourceLine?.box ?? region.box
-  }
-  const byY = (left: PdfPageRegion, right: PdfPageRegion) => {
-    const splitDefinitionOrder = splitNoteDefinitionOrder(left, right)
-    const leftAnchor = readingAnchor(left)
-    const rightAnchor = readingAnchor(right)
-    const inlineOrder = inlineStackedFragmentOrder(
-      left.lines[0] ?? { id: undefined },
-      right.lines[0] ?? { id: undefined },
-    )
-    return (
-      splitDefinitionOrder ??
-      inlineOrder ??
-      (leftAnchor.y - rightAnchor.y ||
-        leftAnchor.x - rightAnchor.x ||
-        left.box.y - right.box.y ||
-        left.box.x - right.box.x ||
-        left.id.localeCompare(right.id))
-    )
-  }
-  return [
-    ...regions.filter((region) => region.column === 'left').sort(byY),
-    ...regions.filter((region) => region.column === 'right').sort(byY),
-    ...regions
-      .filter(
-        (region) => region.column === 'single' || region.column === 'span',
-      )
-      .sort(byY),
-  ]
-}
-
-function orderPageRegions(regions: PdfPageRegion[], layout: ColumnLayout) {
-  const included = regions.filter((region) => region.includedInReadingOrder)
-  const notes = included.filter(
-    (region) => region.kind === 'footnote' || region.kind === 'endnote',
-  )
-  const flow = included.filter(
-    (region) => region.kind !== 'footnote' && region.kind !== 'endnote',
-  )
-  if (layout.split === null) {
-    return restoreInlineStackedAtomicUnits([
-      ...flow.sort(
-        (left, right) =>
-          inlineStackedFragmentOrder(
-            left.lines[0] ?? { id: undefined },
-            right.lines[0] ?? { id: undefined },
-          ) ??
-          (left.box.y - right.box.y || left.box.x - right.box.x),
-      ),
-      ...notes.sort(
-        (left, right) =>
-          splitNoteDefinitionOrder(left, right) ??
-          (left.box.y - right.box.y || left.box.x - right.box.x),
-      ),
-    ])
-  }
-
-  const spanning = flow
-    .filter((region) => region.column === 'span')
-    .sort((left, right) => left.box.y - right.box.y)
-  const columnFlow = flow.filter((region) => region.column !== 'span')
-  const ordered: PdfPageRegion[] = []
-  const emitted = new Set<string>()
-  for (const span of spanning) {
-    const band = columnFlow.filter(
-      (region) => !emitted.has(region.id) && region.box.y < span.box.y,
-    )
-    ordered.push(...columnOrdered(band), span)
-    for (const region of band) emitted.add(region.id)
-  }
-  ordered.push(
-    ...columnOrdered(columnFlow.filter((region) => !emitted.has(region.id))),
-    ...columnOrdered(notes),
-  )
-  return restoreInlineStackedAtomicUnits(ordered)
-}
-
-function edgeEvidence(from: PdfPageRegion, to: PdfPageRegion) {
-  if (from.page !== to.page) {
-    return {
-      code: 'page-sequence' as const,
-      detail: `Page ${from.page} precedes page ${to.page}.`,
-    }
-  }
-  if (to.kind === 'footnote' || to.kind === 'endnote') {
-    return {
-      code: 'note-after-body' as const,
-      detail: 'The separated note region follows the page body flow.',
-    }
-  }
-  if (from.column !== to.column) {
-    return {
-      code: 'column-flow' as const,
-      detail: `${from.column} column flow precedes ${to.column} column flow.`,
-    }
-  }
-  if (
-    from.kind === 'spanning' ||
-    to.kind === 'spanning' ||
-    from.column === 'span' ||
-    to.column === 'span'
-  ) {
-    return {
-      code: 'spanning-boundary' as const,
-      detail: 'A page-spanning region forms a deterministic flow boundary.',
-    }
-  }
-  return {
-    code: 'vertical-flow' as const,
-    detail: 'Regions in the same flow follow normalized vertical geometry.',
-  }
-}
-
-export function hasAcceptedCycle(
-  regionIds: string[],
-  edges: PdfReadingOrderEdge[],
-) {
-  const outgoing = new Map<string, string[]>()
-  for (const edge of edges.filter(
-    (candidate) => candidate.status === 'accepted',
-  )) {
-    const targets = outgoing.get(edge.from) ?? []
-    targets.push(edge.to)
-    outgoing.set(edge.from, targets)
-  }
-  const state = new Map<string, 'visiting' | 'visited'>()
-  for (const root of regionIds) {
-    if (state.has(root)) continue
-    state.set(root, 'visiting')
-    const stack = [{ id: root, nextTarget: 0 }]
-    while (stack.length > 0) {
-      const frame = stack.at(-1)!
-      const targets = outgoing.get(frame.id) ?? []
-      if (frame.nextTarget >= targets.length) {
-        state.set(frame.id, 'visited')
-        stack.pop()
-        continue
-      }
-      const target = targets[frame.nextTarget]
-      frame.nextTarget += 1
-      if (state.get(target) === 'visiting') return true
-      if (state.get(target) === 'visited') continue
-      state.set(target, 'visiting')
-      stack.push({ id: target, nextTarget: 0 })
-    }
-  }
-  return false
-}
-
-function buildReadingOrder(
-  regions: PdfPageRegion[],
-  layouts: Map<number, ColumnLayout>,
-) {
-  const orderedRegions = [...layouts.entries()]
-    .sort(([left], [right]) => left - right)
-    .flatMap(([page, layout]) =>
-      orderPageRegions(
-        regions.filter((region) => region.page === page),
-        layout,
-      ),
-    )
-  const edges: PdfReadingOrderEdge[] = []
-  for (let index = 0; index < orderedRegions.length - 1; index += 1) {
-    const from = orderedRegions[index]
-    const to = orderedRegions[index + 1]
-    const layout = layouts.get(from.page)
-    const crossColumnBoundary =
-      from.page === to.page && from.column === 'left' && to.column === 'right'
-    const ambiguousBoundary = crossColumnBoundary && Boolean(layout?.ambiguous)
-    const resolvedBoundary =
-      crossColumnBoundary && layout?.resolution?.status === 'resolved'
-    edges.push({
-      id: `reading-edge-${String(edges.length + 1).padStart(4, '0')}`,
-      from: from.id,
-      to: to.id,
-      status: ambiguousBoundary ? 'candidate' : 'accepted',
-      confidence:
-        ambiguousBoundary || resolvedBoundary
-          ? (layout?.resolution?.confidence ?? 0.5)
-          : 0.96,
-      evidence: ambiguousBoundary
-        ? [
-            ...(layout?.resolution?.evidence ?? []),
-            {
-              code: 'ambiguous-column-flow',
-              detail: `Both column orders remain candidates because confidence ${layout?.resolution?.confidence ?? 0.5} is below threshold ${READING_ORDER_RESOLUTION_THRESHOLD}.`,
-            },
-          ]
-        : resolvedBoundary
-          ? [...layout.resolution!.evidence, edgeEvidence(from, to)]
-          : [edgeEvidence(from, to)],
-      sourceBoxes: [from.box, to.box],
-    })
-    if (ambiguousBoundary) {
-      const rightRegions = orderedRegions.filter(
-        (region) => region.page === from.page && region.column === 'right',
-      )
-      const leftRegions = orderedRegions.filter(
-        (region) => region.page === from.page && region.column === 'left',
-      )
-      const reverseFrom = rightRegions.at(-1)
-      const reverseTo = leftRegions[0]
-      if (reverseFrom && reverseTo) {
-        edges.push({
-          id: `reading-edge-${String(edges.length + 1).padStart(4, '0')}`,
-          from: reverseFrom.id,
-          to: reverseTo.id,
-          status: 'candidate',
-          confidence: layout?.resolution?.confidence ?? 0.5,
-          evidence: [
-            ...(layout?.resolution?.evidence ?? []),
-            {
-              code: 'ambiguous-column-flow',
-              detail: `The reverse column order is retained for bounded review below threshold ${READING_ORDER_RESOLUTION_THRESHOLD}.`,
-            },
-          ],
-          sourceBoxes: [reverseFrom.box, reverseTo.box],
-        })
-      }
-    }
-  }
-  const regionIds = regions.map((region) => region.id)
-  const resolutions = [...layouts.entries()]
-    .filter(
-      (
-        entry,
-      ): entry is [
-        number,
-        ColumnLayout & { resolution: NonNullable<ColumnLayout['resolution']> },
-      ] => entry[1].resolution !== null,
-    )
-    .map<PdfReadingOrderResolution>(([page, layout]) => ({
-      page,
-      ...layout.resolution,
-      regionIds: regions
-        .filter(
-          (region) => region.page === page && region.includedInReadingOrder,
-        )
-        .map((region) => region.id),
-    }))
-  const acyclic = !hasAcceptedCycle(regionIds, edges)
-  const unresolvedEdgeCount = edges.filter(
-    (edge) => edge.status === 'candidate',
-  ).length
-  const evaluation: PdfReadingOrderEvaluation = {
-    schemaVersion: '1.0.0',
-    algorithm: 'deterministic-geometry-v1',
-    mode: 'deterministic-only',
-    regionCount: orderedRegions.length,
-    acceptedEdgeCount: edges.length - unresolvedEdgeCount,
-    unresolvedEdgeCount,
-    cycleRate: acyclic ? 0 : 1,
-    orderAccuracy: null,
-    provider: null,
-    modelVersion: null,
-    latencyMs: 0,
-    costUsd: 0,
-    reviewRequired: !acyclic || unresolvedEdgeCount > 0,
-  }
-  return {
-    schemaVersion: '1.0.0',
-    regionIds,
-    order: orderedRegions.map((region) => region.id),
-    edges,
-    resolutions,
-    acyclic,
-    evaluation,
-  } satisfies PdfReadingOrderGraph
-}
-
-export function evaluateReadingOrder(
-  graph: PdfReadingOrderGraph,
-  expectedOrder?: readonly string[],
-) {
-  if (!expectedOrder) return { ...graph.evaluation }
-  const expectedPositions = new Map(
-    expectedOrder.map((id, index) => [id, index]),
-  )
-  const comparable = graph.order.filter((id) => expectedPositions.has(id))
-  const correct = comparable.filter(
-    (id, index) => expectedPositions.get(id) === index,
-  ).length
-  return {
-    ...graph.evaluation,
-    orderAccuracy:
-      expectedOrder.length === 0
-        ? 1
-        : rounded(correct / Math.max(expectedOrder.length, graph.order.length)),
-  }
 }
 
 export function reconstructPageRegions(
